refactor(ui): extract non-chat component groups
