+++ apps/desktop/src/main/__tests__/chat-tool-card-cascade-contract.test.ts
@@ -122,23 +122,23 @@ describe('chat tool-card migration contract (#332 PR3b)', () => {
 
   it('keeps the computed-style fixture covering every production tool status', async () => {
     // The zero-visual proof is only as complete as the statuses it renders. The
-    // truth source is components.tsx's `STATUS_LABEL` (a `Record` over the full
+    // truth source is tool-activity.tsx's `STATUS_LABEL` (a `Record` over the full
     // `ToolActivityItem['status']` union — its keys ARE every production status).
     // The harness must render a card for each: a new status that escapes the
     // fixture would have NO diffed row, so its tint could drift unseen (the
     // `pending` gap this guard was added to close). No card-level exclusions —
     // even `running` renders a card; only its animated DOT id is left out of IDS.
     const componentsSrc = await readFile(
-      resolve(REPO_ROOT, 'packages', 'ui', 'src', 'components.tsx'),
+      resolve(REPO_ROOT, 'packages', 'ui', 'src', 'tool-activity.tsx'),
       'utf8',
     );
     // Anchor on the FULL `ToolActivityItem` signature — a bare `const STATUS_LABEL`
     // prefix-matches the unrelated session-status `STATUS_LABEL_BY_STATUS` map.
     const labelStart = componentsSrc.indexOf("const STATUS_LABEL: Record<ToolActivityItem['status']");
-    assert.ok(labelStart !== -1, 'failed to locate the tool STATUS_LABEL declaration in components.tsx');
+    assert.ok(labelStart !== -1, 'failed to locate the tool STATUS_LABEL declaration in tool-activity.tsx');
     const labelBlock = componentsSrc.slice(labelStart, componentsSrc.indexOf('};', labelStart));
     const statuses = [...labelBlock.matchAll(/^\s*(\w+):/gm)].map((m) => m[1]);
-    assert.ok(statuses.length >= 5, 'failed to parse STATUS_LABEL keys from components.tsx');
+    assert.ok(statuses.length >= 5, 'failed to parse STATUS_LABEL keys from tool-activity.tsx');
 
     const harnessSrc = await readFile(
       resolve(REPO_ROOT, 'scripts', 'check-chat-marker-computed-style.mjs'),
