+++ apps/desktop/src/main/__tests__/overlay-scrollbars-contract.test.ts
@@ -61,10 +61,11 @@ describe('OverlayScrollbars integration contract', () => {
 
   it('migrates the primary app scroll surfaces onto OverlayScrollArea', async () => {
     const components = await repoFile('packages/ui/src/components.tsx');
+    const sessionListPanel = await repoFile('packages/ui/src/session-list-panel.tsx');
     const settings = await repoFile('apps/desktop/src/renderer/settings/SettingsModal.tsx');
 
-    assert.match(components, /import \{ OverlayScrollArea \} from '\.\/overlay-scroll-area\.js';/, 'main UI components must import OverlayScrollArea');
-    assert.match(components, /<OverlayScrollArea[\s\S]*className="maka-list-stack"[\s\S]*contentClassName="maka-list-stackContent"/, 'sidebar session list must use OverlayScrollArea');
+    assert.match(sessionListPanel, /import \{ OverlayScrollArea \} from '\.\/overlay-scroll-area\.js';/, 'session list panel must import OverlayScrollArea');
+    assert.match(sessionListPanel, /<OverlayScrollArea[\s\S]*className="maka-list-stack"[\s\S]*contentClassName="maka-list-stackContent"/, 'sidebar session list must use OverlayScrollArea');
     assert.match(components, /<OverlayScrollArea[\s\S]*ref=\{scrollRef\}[\s\S]*className="maka-chat messages"[\s\S]*onScroll=\{onScroll\}/, 'active chat message list must keep its onScroll viewport handler on OverlayScrollArea');
     assert.match(settings, /OverlayScrollArea/, 'Settings content pane must use OverlayScrollArea');
     assert.match(settings, /<OverlayScrollArea[\s\S]*className="settingsPageContent"[\s\S]*contentClassName="settingsPageContentInner"/, 'Settings content pane must preserve its layout classes through OverlayScrollArea');
