+++ apps/desktop/src/main/__tests__/tool-args-redaction-contract.test.ts
@@ -18,9 +18,12 @@ describe('tool and permission args redaction', () => {
   });
 
   it('routes ToolActivity and PermissionDialog args through formatRedactedJson', async () => {
-    const source = await readFile(join(process.cwd(), '../../packages/ui/src/components.tsx'), 'utf8');
-    const toolActivity = source.match(/export function ToolActivity[\s\S]*?function ToolOutputStream/)?.[0] ?? '';
-    const permissionDialog = source.match(/export function PermissionDialog[\s\S]*?function renderPermissionSummary/)?.[0] ?? '';
+    const [toolSource, permissionSource] = await Promise.all([
+      readFile(join(process.cwd(), '../../packages/ui/src/tool-activity.tsx'), 'utf8'),
+      readFile(join(process.cwd(), '../../packages/ui/src/permission-dialog.tsx'), 'utf8'),
+    ]);
+    const toolActivity = toolSource.match(/export function ToolActivity[\s\S]*?function ToolOutputStream/)?.[0] ?? '';
+    const permissionDialog = permissionSource.match(/export function PermissionDialog[\s\S]*?function renderPermissionSummary/)?.[0] ?? '';
 
     assert.match(toolActivity, /\{formatRedactedJson\(item\.args\)\}/);
     assert.doesNotMatch(toolActivity, /JSON\.stringify\(item\.args/);
@@ -37,14 +40,14 @@ describe('tool and permission args redaction', () => {
     assert.match(rendered, /Authorization: Bearer/);
     assert.ok(rendered.length <= 241);
 
-    const source = await readFile(join(process.cwd(), '../../packages/ui/src/components.tsx'), 'utf8');
+    const source = await readFile(join(process.cwd(), '../../packages/ui/src/tool-activity.tsx'), 'utf8');
     const toolActivity = source.match(/export function ToolActivity[\s\S]*?function ToolOutputStream/)?.[0] ?? '';
     assert.match(toolActivity, /\{formatToolIntent\(item\.intent\)\}/);
     assert.doesNotMatch(toolActivity, /\{item\.intent\}/);
   });
 
   it('redacts permission summary previews before rendering command, path, or file content', async () => {
-    const source = await readFile(join(process.cwd(), '../../packages/ui/src/components.tsx'), 'utf8');
+    const source = await readFile(join(process.cwd(), '../../packages/ui/src/permission-dialog.tsx'), 'utf8');
     const summary = source.match(/function renderPermissionSummary[\s\S]*?function permissionValuePreview/)?.[0] ?? '';
 
     assert.match(summary, /\{redactSecrets\(command\)\}/);
