+++ apps/desktop/src/main/__tests__/visible-copy-hygiene-contract.test.ts
@@ -287,7 +287,7 @@ describe('visible-copy hygiene contract (PR-SIDEBAR-IA-0 Phase 3 P0 fixup v2)', 
 
 describe('terminal truncation handoff contract', () => {
   it('shows a deep-research handoff when terminal output is capped', async () => {
-    const componentsPath = resolve(process.cwd(), '..', '..', 'packages', 'ui', 'src', 'components.tsx');
+    const componentsPath = resolve(process.cwd(), '..', '..', 'packages', 'ui', 'src', 'tool-activity.tsx');
     const src = await readFile(componentsPath, 'utf8');
 
     assert.match(
@@ -531,7 +531,7 @@ describe('turn footer copy feedback contract', () => {
 
 describe('tool error copy feedback contract', () => {
   it('routes tool-error copy through the shared guarded feedback path', async () => {
-    const componentsPath = resolve(process.cwd(), '..', '..', 'packages', 'ui', 'src', 'components.tsx');
+    const componentsPath = resolve(process.cwd(), '..', '..', 'packages', 'ui', 'src', 'tool-activity.tsx');
     const src = await readFile(componentsPath, 'utf8');
     const block = src.match(/function ToolErrorBanner[\s\S]*?export function OverlayHost/)?.[0] ?? '';
 
@@ -562,7 +562,7 @@ describe('tool error copy feedback contract', () => {
     // scan would false-pass if the string drifted to another component. These are
     // arbitrary-value utilities (source == computed), so this source contract is the
     // proof; the computed-style harness only re-diffs the non-trivial container box.
-    const componentsPath = resolve(process.cwd(), '..', '..', 'packages', 'ui', 'src', 'components.tsx');
+    const componentsPath = resolve(process.cwd(), '..', '..', 'packages', 'ui', 'src', 'tool-activity.tsx');
     const src = await readFile(componentsPath, 'utf8');
     const block = src.match(/function ToolErrorBanner[\s\S]*?export function OverlayHost/)?.[0] ?? '';
