+++ packages/ui/src/components.tsx
@@ -200,648 +200,49 @@ import { Empty, EmptyContent, EmptyDescription, EmptyHeader, EmptyMedia, EmptyTi
 import { InputGroup, InputGroupAddon, InputGroupInput } from './primitives/input-group.js';
 import { Kbd } from './primitives/kbd.js';
 import { Menu, MenuItem, MenuPopup, MenuTrigger } from './primitives/menu.js';
+import type { NavSelection } from './nav-selection.js';
+import { EmptyState } from './empty-state.js';
+import { CapabilityAuditStrip } from './capability-audit-strip.js';
+import type {
+  DailyReviewBridge,
+  DailyReviewMarkdownActionInput,
+  PlanReminderDraftInput,
+  PlanReminderUpdatePatch,
+  SkillEntry,
+} from './module-panel-types.js';
+import { ToolActivity } from './tool-activity.js';
 
-export type NavSelection =
-  | { section: 'sessions'; filter: SessionFilter }
-  | { section: 'automations' }
-  | { section: 'skills' }
-  | { section: 'daily-review' };
+export type { NavSelection } from './nav-selection.js';
+export { useModalA11y } from './modal-a11y.js';
+export { CapabilityAuditStrip } from './capability-audit-strip.js';
+export { SearchModal } from './search-modal.js';
+export { SessionListPanel } from './session-list-panel.js';
+export type { SkillEntry } from './module-panel-types.js';
+export { describeLoadToolResult, formatRedactedJson, formatToolIntent, loadToolDisplayName } from './tool-format.js';
+export { formatBytes, OverlayHost, ToolActivity } from './tool-activity.js';
+export { PermissionDialog } from './permission-dialog.js';
 
-type SessionFilter = 'chats' | 'flagged' | 'archived';
 
-/**
- * Sidebar module ids. "sessions" is still the lower history region label,
- * but it is no longer rendered as a top-level nav row: "新任务" creates
- * the chat/task, while the history list below shows prior sessions.
- *
- * Includes `search` even though it is not a `NavSelection.section` —
- * search is a modal trigger and needs a label/icon but no section.
- */
-type ModuleNavId = NavSelection['section'] | 'search';
 
-/**
- * Top-level module nav labels. Chinese-first per xuan `47e204f2` #5.
- * Keyed by `ModuleNavId` so the `search` modal trigger gets a label
- * even though it is not a `NavSelection.section`.
- */
-const MODULE_NAV_LABEL: Record<ModuleNavId, string> = {
-  sessions: '会话',
-  search: '搜索',
-  automations: '定时任务',
-  skills: '技能',
-  'daily-review': '每日回顾',
-};
 
-/**
- * Hook for accessible modal dialogs.
- *
- * - Saves the element that had focus before the modal opened.
- * - Moves focus to the first focusable element inside the modal on mount
- *   (or the container itself if no focusable child exists).
- * - Traps Tab/Shift+Tab inside the modal.
- * - Optionally closes the modal on Escape.
- * - Restores focus to the previously-focused element on unmount.
- *
- * Implements rule "3. focus and dialogs (critical)" from the
- * fixing-accessibility skill.
- */
-export function useModalA11y(
-  containerRef: RefObject<HTMLElement | null>,
-  onEscape?: () => void,
-  initialFocusRef?: RefObject<HTMLElement | null>,
-): void {
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
 
-    const previouslyFocused = document.activeElement as HTMLElement | null;
 
-    const preferredInitial = initialFocusRef?.current;
-    const initial = preferredInitial && container.contains(preferredInitial)
-      ? preferredInitial
-      : getFocusable(container)[0];
-    if (initial) {
-      initial.focus({ preventScroll: true });
-    } else {
-      if (!container.hasAttribute('tabindex')) container.setAttribute('tabindex', '-1');
-      container.focus({ preventScroll: true });
-    }
 
-    function onKeyDown(event: globalThis.KeyboardEvent) {
-      if (!container) return;
-      if (event.key === 'Escape' && onEscape) {
-        event.stopPropagation();
-        event.preventDefault();
-        onEscape();
-        return;
-      }
-      if (event.key !== 'Tab') return;
-      const items = getFocusable(container);
-      if (items.length === 0) {
-        event.preventDefault();
-        return;
-      }
-      const first = items[0]!;
-      const last = items[items.length - 1]!;
-      const active = document.activeElement;
-      if (event.shiftKey && (active === first || !container.contains(active))) {
-        event.preventDefault();
-        last.focus({ preventScroll: true });
-      } else if (!event.shiftKey && (active === last || !container.contains(active))) {
-        event.preventDefault();
-        first.focus({ preventScroll: true });
-      }
-    }
 
-    container.addEventListener('keydown', onKeyDown);
-    return () => {
-      container.removeEventListener('keydown', onKeyDown);
-      // Defer restoration so any in-flight focus changes (e.g. clicking a
-      // button that unmounts the modal) settle before we yank focus back.
-      queueMicrotask(() => {
-        if (document.contains(container)) return;
-        if (previouslyFocused && document.contains(previouslyFocused)) {
-          previouslyFocused.focus?.({ preventScroll: true });
-        }
-      });
-    };
-  }, [containerRef, onEscape, initialFocusRef]);
-}
 
-const FOCUSABLE_SELECTOR =
-  'a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), iframe, [tabindex]:not([tabindex="-1"])';
-
-function getFocusable(container: HTMLElement): HTMLElement[] {
-  return Array.from(container.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR)).filter(
-    (element) => !element.hasAttribute('inert') && isVisible(element),
-  );
-}
-
-function isVisible(element: HTMLElement): boolean {
-  if (element.hidden) return false;
-  // offsetParent is null for display:none ancestors and fixed-positioned roots,
-  // but our modal elements are always rendered visible — so this is a sufficient
-  // approximation without forcing layout.
-  return element.offsetParent !== null || element === document.activeElement;
-}
 
 function Count(props: { value: number }) {
   if (props.value <= 0) return null;
   return <small>{props.value}</small>;
 }
 
-export interface SkillEntry {
-  id: string;
-  name: string;
-  description: string;
-  path: string;
-  /**
-   * Tools the skill *declares* it would like to use. This is a request, not
-   * a grant — PermissionEngine still applies. We surface the list so users
-   * can see what a skill is asking for before they install / enable it.
-   */
-  declaredTools?: string[];
-}
 
-type PlanReminderDraftInput = {
-  title: string;
-  note?: string;
-  runAt: number;
-  recurrence?: PlanReminderRecurrence;
-  cronExpression?: string;
-  delivery?: PlanReminderDeliveryTarget;
-};
 
-type PlanReminderUpdatePatch = {
-  title?: string;
-  note?: string;
-  runAt?: number;
-  recurrence?: PlanReminderRecurrence;
-  cronExpression?: string;
-  delivery?: PlanReminderDeliveryTarget;
-  enabled?: boolean;
-};
 
-type SessionRowActionId = 'flag' | 'archive' | 'rename' | 'delete';
 
-interface SessionRowActions {
-  /** Flag (pin) state toggle. */
-  onToggleFlag(sessionId: string, next: boolean): void | Promise<void>;
-  /** Move to / out of the archive bucket. */
-  onArchive(sessionId: string): void | Promise<void>;
-  onUnarchive(sessionId: string): void | Promise<void>;
-  /** Rename via inline prompt. Receives the new (trimmed) name. */
-  onRename(sessionId: string, name: string): void | Promise<void>;
-  /** Permanent removal — caller is responsible for the confirm gate. */
-  onDelete(sessionId: string): void | Promise<void>;
-}
 
-export function SessionListPanel(props: {
-  selection: NavSelection;
-  sessionCounts: Record<SessionFilter, number>;
-  sessions: SessionSummary[];
-  activeId?: string;
-  skills?: SkillEntry[];
-  onRefreshSkills?(): void | Promise<void>;
-  onCreateSkillTemplate?(): void | Promise<void>;
-  planReminders?: PlanReminder[];
-  /**
-   * Per-session-id boolean flag: true when the session has a live streaming
-   * delta in flight. Rendered as a small pulsing accent dot on the row.
-   * Caller (main.tsx) derives this from `streamingBySession` so the sidebar
-   * shows live activity without subscribing to the stream itself.
-   */
-  streamingSessionIds?: Set<string>;
-  /**
-   * Per-session-id boolean flag: true when the session's backend / connection
-   * is stale (`backend='fake'` or `llmConnectionSlug` no longer resolves).
-   * The row dims + shows a small "已过期" pill so users notice in the list
-   * before clicking in and seeing the chat header banner. Caller derives this
-   * by joining `sessions` against `connections` — keeps SessionListPanel
-   * unaware of the connection store.
-   */
-  staleSessionIds?: Set<string>;
-  /**
-   * Pre-computed status-driven groups for the session list (PR109b).
-   * When provided, replaces the date-bucket grouping for the `chats`
-   * filter. Caller derives this via `deriveSessionStatusGroups()` from
-   * `apps/desktop/src/renderer/session-status-grouping.ts`. Each group
-   * carries its own collapsible/defaultExpanded flag so the panel
-   * doesn't have to know about Archived being closed by default.
-   */
-  statusGroups?: ReadonlyArray<{
-    id: string;
-    label: string;
-    sessions: SessionSummary[];
-    collapsible: boolean;
-    defaultExpanded: boolean;
-  }>;
-  onSelectSession(sessionId: string): void;
-  onSelect(selection: NavSelection): void;
-  onOpenSettings(): void;
-  userLabel?: string;
-  onNew(): void;
-  onOpenSkill?(skillId: string): void;
-  /**
-   * PR-SIDEBAR-IA-0 Phase 2 fixup (xuan `91401163` + `94c7bf0f`):
-   * Sidebar `搜索` nav row click handler. Opens a dedicated Search
-   * modal hosted by the application shell; does NOT change
-   * `selection`. The shell owns the real search backend and modal
-   * lifecycle behind this callback.
-   */
-  onOpenSearchModal?(): void;
-  onRefreshPlanReminders?(): void | Promise<void>;
-  onCreatePlanReminder?(input: PlanReminderDraftInput): boolean | Promise<boolean> | void | Promise<void>;
-  onUpdatePlanReminder?(id: string, patch: PlanReminderUpdatePatch): boolean | Promise<boolean> | void | Promise<void>;
-  onTogglePlanReminder?(id: string, enabled: boolean): void | Promise<void>;
-  onTriggerPlanReminderNow?(id: string): void | Promise<void>;
-  onSnoozePlanReminder?(id: string): void | Promise<void>;
-  onClearPlanReminderRunHistory?(id: string): void | Promise<void>;
-  onDeletePlanReminder?(id: string): void | Promise<void>;
-  onCopyDailyReviewMarkdown?(input: DailyReviewMarkdownActionInput): Promise<void> | void;
-  onSaveDailyReviewMarkdown?(input: DailyReviewMarkdownActionInput): Promise<void> | void;
-  /**
-   * PR-DAILY-REVIEW-MVP-0: bridge for the `每日回顾` panel. When
-   * provided, the daily-review section renders the real panel instead
-   * of the fallback view. When `undefined` (e.g. in visual-smoke
-   * fixtures without an IPC layer), it falls back to an explicit
-   * bridge-missing state.
-   */
-  dailyReviewBridge?: DailyReviewBridge;
-  rowActions?: SessionRowActions;
-  sidebarCollapsed?: boolean;
-  onToggleSidebar?(): void;
-}) {
-  // 参考实现 keeps the lower sidebar region as stable chat history
-  // even when Skills / Scheduled Tasks are open in the main pane.
-  const sessionListTitle = MODULE_NAV_LABEL.sessions;
-  // PR-UX-POLISH-1 commit 4 (WAWQAQ msg `e0dbad11` + kenji msg
-  // `2844f64f`): in-list `筛选会话` filter input removed. All search
-  // capability lives in the top-level `搜索` modal (PR-SEARCH-MODAL-
-  // REAL-0 wires it to `window.maka.search.thread()` in the same PR).
-  // The previous `searchQuery` state + `searchInputRef` + ⌘F/Ctrl+F
-  // focus binding are gone with it; ⌘F is freed for future use.
-  // `filteredSessions` collapses to a direct passthrough of
-  // `props.sessions` — group rendering downstream still partitions
-  // by status / time / filter.
-  const filteredSessions = props.sessions;
 
-  function handleListKeyDown(event: KeyboardEvent<HTMLDivElement>) {
-    // PR-SIDEBAR-IA-0 Phase 2 fixup (xuan `71687cc7`): the
-    // ArrowLeft/ArrowRight filter cycle was REMOVED. Hidden state
-    // without visible UI is harder for users to discover and harder
-    // for review to verify. If we re-introduce Pinned/Archived
-    // access in the future it will be a deliberate, visible,
-    // lightweight control (per kenji `9f683ea8`).
-    if (event.key === 'Delete' || event.key === 'Backspace') {
-      // Delete on a focused row opens the App-level confirmation (which
-      // toast.confirm()s); we do not delete silently per the lifecycle
-      // contract.
-      const active = document.activeElement as HTMLElement | null;
-      const row = active?.closest('.maka-list-row');
-      const sessionId = row?.querySelector<HTMLButtonElement>('.maka-list-row-main')?.dataset.sessionId;
-      if (sessionId && props.rowActions) {
-        event.preventDefault();
-        void props.rowActions.onDelete(sessionId);
-      }
-      return;
-    }
-    if (event.key !== 'ArrowUp' && event.key !== 'ArrowDown' && event.key !== 'Home' && event.key !== 'End') {
-      return;
-    }
-    const list = event.currentTarget;
-    const focusables = Array.from(
-      list.querySelectorAll<HTMLButtonElement>('.maka-list-row-main'),
-    );
-    if (focusables.length === 0) return;
-    const active = document.activeElement as HTMLElement | null;
-    const currentIndex = active ? focusables.indexOf(active as HTMLButtonElement) : -1;
-    let nextIndex = currentIndex;
-    switch (event.key) {
-      case 'ArrowDown':
-        nextIndex = currentIndex < 0 ? 0 : Math.min(focusables.length - 1, currentIndex + 1);
-        break;
-      case 'ArrowUp':
-        nextIndex = currentIndex <= 0 ? 0 : currentIndex - 1;
-        break;
-      case 'Home':
-        nextIndex = 0;
-        break;
-      case 'End':
-        nextIndex = focusables.length - 1;
-        break;
-    }
-    if (nextIndex === currentIndex) return;
-    event.preventDefault();
-    focusables[nextIndex]?.focus({ preventScroll: false });
-    focusables[nextIndex]?.scrollIntoView({ block: 'nearest' });
-  }
 
-  // PR-PARCHMENT-HOME-9 (WAWQAQ msg `781852eb`): restored module nav
-  // helpers. `search` is a transient modal trigger and never
-  // "active". Plan count shows unread reminders as a small chip.
-  const isModuleActive = (id: ModuleNavId) => {
-    if (id === 'search') return false;
-    return props.selection.section === id;
-  };
-  const activePlanReminderCount = (props.planReminders ?? [])
-    .filter((reminder) => reminder.status !== 'completed')
-    .length;
-  function selectModule(id: ModuleNavId) {
-    if (id === 'search') {
-      props.onOpenSearchModal?.();
-      return;
-    }
-    if (id === 'sessions') {
-      props.onSelect({ section: 'sessions', filter: 'chats' });
-      return;
-    }
-    if (id === 'automations') props.onSelect({ section: 'automations' });
-    else if (id === 'skills') props.onSelect({ section: 'skills' });
-    else if (id === 'daily-review') props.onSelect({ section: 'daily-review' });
-  }
-
-  return (
-    <aside
-      className="maka-session-panel agents-sidebar"
-      aria-label="对话列表"
-      data-collapsed={props.sidebarCollapsed ? 'true' : undefined}
-    >
-      <header className="maka-session-panel-header">
-        <div className="maka-sidebar-drag-strip">
-          {/* PR-SIDEBAR-HEADER-BUTTONS-PRIMITIVE-0 (round 5/30):
-              both icon affordances in the sidebar drag strip now
-              flow through UiButton. variant="quiet" + size="icon-sm"
-              match the 24×24 quiet-icon shape; the custom class
-              still owns -webkit-app-region + the bespoke
-              focus-visible ring (3px accent shadow). */}
-          {/* PR-SIDEBAR-WORDMARK-KILL-0 (WAWQAQ msg `1fae5521` 2026-06-24
-              "现在侧边栏左上角有Maka这个单词，好丑啊"): dropped the
-              Georgia serif `Maka` wordmark that PR #190 added next
-              to the traffic lights. The app icon + window title bar
-              already carry the brand identity; the wordmark inside
-              the sidebar drag strip read as redundant chrome. */}
-          {props.onOpenSearchModal && (
-            <UiButton
-              className="maka-sidebar-search-button"
-              variant="quiet"
-              size="icon-sm"
-              type="button"
-              data-maka-search-trigger="true"
-              onClick={props.onOpenSearchModal}
-              aria-label="搜索对话"
-              title="搜索对话"
-            >
-              <Search size={16} strokeWidth={1.65} aria-hidden="true" />
-            </UiButton>
-          )}
-          <UiButton
-            className="maka-sidebar-toggle"
-            variant="quiet"
-            size="icon-sm"
-            type="button"
-            onClick={props.onToggleSidebar}
-            aria-label={props.sidebarCollapsed ? '展开侧边栏' : '收起侧边栏'}
-            aria-expanded={!props.sidebarCollapsed}
-            title={props.sidebarCollapsed ? '展开侧边栏' : '收起侧边栏'}
-          >
-            {props.sidebarCollapsed ? (
-              <PanelLeftOpen size={16} strokeWidth={1.65} aria-hidden="true" />
-            ) : (
-              <PanelLeftClose size={16} strokeWidth={1.65} aria-hidden="true" />
-            )}
-          </UiButton>
-        </div>
-      </header>
-
-      {/*
-        内部参考式 IA: the primary rail is a flat list of actions/modules.
-        "会话" is the lower history region, not a second top-level page entry.
-      */}
-      {/* PR-SIDEBAR-NAV-ROWS-PRIMITIVE-0 (round 13/30): the 4
-          main nav rows (新任务 / 每日回顾 / 技能 / 定时任务) were
-          all raw <button>. Routed through UiButton with the
-          newly-introduced size="nav" variant — `size="nav"`
-          contributes no layout utilities so .maka-nav-row's
-          tight 30px min-height + 3px 6px padding + 14px font +
-          grid-template-columns stays the source of truth. The
-          primitive contributes `:active scale` + focus-visible
-          + disabled-state contract uniformly with the rest of
-          the panel. */}
-      <nav className="maka-sidebar-modules" aria-label="主导航">
-        <UiButton
-          variant="quiet"
-          size="nav"
-          className="maka-nav-row maka-nav-new-task"
-          aria-label="新任务"
-          type="button"
-          onClick={props.onNew}
-        >
-          <SquarePen className="maka-nav-icon" strokeWidth={1.5} aria-hidden="true" />
-          <span>新任务</span>
-        </UiButton>
-        {/* PR-UI-PIXEL-5 (2026-06-21): the standalone 搜索 nav row was
-            removed — search is reachable from the dedicated search button in
-            the sidebar header (`.maka-sidebar-search-button`) and ⌘K, so a
-            second entry point was redundant. The `search` module id + label
-            are kept for those triggers. */}
-        <UiButton
-          variant="quiet"
-          size="nav"
-          className="maka-nav-row"
-          data-active={isModuleActive('daily-review')}
-          aria-current={isModuleActive('daily-review') ? 'page' : undefined}
-          aria-label={MODULE_NAV_LABEL['daily-review']}
-          type="button"
-          onClick={() => selectModule('daily-review')}
-        >
-          <LineChart className="maka-nav-icon" strokeWidth={1.5} aria-hidden="true" />
-          <span>{MODULE_NAV_LABEL['daily-review']}</span>
-        </UiButton>
-        <UiButton
-          variant="quiet"
-          size="nav"
-          className="maka-nav-row"
-          data-active={isModuleActive('skills')}
-          aria-current={isModuleActive('skills') ? 'page' : undefined}
-          aria-label={MODULE_NAV_LABEL.skills}
-          type="button"
-          onClick={() => selectModule('skills')}
-        >
-          <Sparkles className="maka-nav-icon" strokeWidth={1.5} aria-hidden="true" />
-          <span>{MODULE_NAV_LABEL.skills}</span>
-        </UiButton>
-        <UiButton
-          variant="quiet"
-          size="nav"
-          className="maka-nav-row"
-          data-active={isModuleActive('automations')}
-          aria-current={isModuleActive('automations') ? 'page' : undefined}
-          type="button"
-          onClick={() => selectModule('automations')}
-          aria-label={activePlanReminderCount > 0 ? `定时任务，${activePlanReminderCount} 个未完成提醒` : MODULE_NAV_LABEL.automations}
-        >
-          <Clock className="maka-nav-icon" strokeWidth={1.5} aria-hidden="true" />
-          <span>{MODULE_NAV_LABEL.automations}</span>
-          {activePlanReminderCount > 0 && (
-            <small className="maka-nav-count" aria-hidden="true">{activePlanReminderCount}</small>
-          )}
-        </UiButton>
-      </nav>
-
-      {/*
-        PR-UX-POLISH-1 commit 4 (WAWQAQ msg `e0dbad11` + kenji msg
-        `2844f64f` blocker #2): the in-list `筛选会话` filter input
-        is REMOVED entirely. Search capability lives only in the
-        top-level `搜索` modal (Cmd-click / sidebar nav → modal),
-        which the same PR wires to real `window.maka.search.thread()`
-        backend. No more duplicated search affordances; one canonical
-        entry point.
-
-        Removed with it:
-        - `searchQuery` state + `searchInputRef`
-        - `useMemo(() => filter sessions by name)` — sessions pass
-          through unchanged; group rendering still partitions.
-        - `useEffect(⌘F focuses input)` — ⌘F is freed.
-        - `.maka-session-search` JSX block + clear button.
-        - "没有匹配的会话" empty state (the only consumer was
-          `filteredSessions.length === 0 && searchQuery.length > 0`).
-      */}
-
-      <section className="maka-session-list" aria-label={sessionListTitle}>
-        {props.sessions.length === 0 ? (
-          // WAWQAQ msg `f56f38c1` (2026-06-20): the create-session CTA
-          // belongs in the sidebar header / nav rail, never in the
-          // bottom session-history empty state. The empty state here is
-          // pure "no sessions yet" copy — no inline CTA. The top-of-
-          // sidebar `+ 新任务` button is the only create-session entry.
-          <EmptyState
-            Icon={MessageSquare}
-            title="等待开始对话"
-            body="和 Maka 的对话会出现在这里。"
-            extraClassName="maka-session-empty-state"
-          />
-        ) : (
-          <OverlayScrollArea
-            className="maka-list-stack"
-            viewportClassName="maka-list-stackViewport"
-            contentClassName="maka-list-stackContent"
-            onKeyDown={handleListKeyDown}
-          >
-            <SessionListGroups
-              groups={
-                props.statusGroups
-                  ? props.statusGroups.map((g) => ({
-                      key: g.id,
-                      label: g.label,
-                      sessions: g.sessions,
-                      collapsible: g.collapsible,
-                      defaultExpanded: g.defaultExpanded,
-                    }))
-                  : groupSessionsForFilter(filteredSessions, { section: 'sessions', filter: 'chats' }).map((g) => ({
-                      key: g.label,
-                      label: g.label,
-                      sessions: g.sessions,
-                      collapsible: false,
-                      defaultExpanded: true,
-                    }))
-              }
-              activeId={props.activeId}
-              streamingSessionIds={props.streamingSessionIds}
-              staleSessionIds={props.staleSessionIds}
-              onSelectSession={props.onSelectSession}
-              rowActions={props.rowActions}
-            />
-          </OverlayScrollArea>
-        )}
-      </section>
-
-      <footer className="maka-session-panel-footer">
-        {/* Maka has no account system — the sidebar footer is a thin
-            settings affordance only. The earlier `.maka-sidebar-account`
-            "Free Plan" widget falsely implied a subscription model and
-            was removed per WAWQAQ msg cad3dec4. About / version info
-            still reachable via Settings → 关于. */}
-        {/* PR-SIDEBAR-SETTINGS-BUTTON-PRIMITIVE-0 (round 4/30):
-            was a raw <button>. Route through UiButton so the
-            sidebar footer respects the same :active scale,
-            hover token, focus-visible, and `data-pressed`
-            contract as every other interactive element in
-            this panel. The custom class still owns the
-            full-width grid layout (24px icon + 1fr label). */}
-        <UiButton
-          className="maka-sidebar-settings-button"
-          variant="quiet"
-          type="button"
-          onClick={props.onOpenSettings}
-          aria-label="设置"
-          title="设置"
-        >
-          <Settings className="maka-nav-icon" strokeWidth={1.5} aria-hidden="true" />
-          <span>设置</span>
-        </UiButton>
-        {/*
-          PR-UX-POLISH-1 commit 4 (WAWQAQ msg `e0dbad11` + kenji
-          msg `2844f64f` blocker #1): the `? 快捷键` chip in the
-          sidebar footer is removed. The sidebar footer is for
-          product nav/state, not help affordances. Keyboard
-          shortcut discoverability moves to Command Palette
-          (`查看快捷键` entry) and the existing global `?` keydown
-          listener stays — power users still hit `?` to open the
-          modal; new users find it via Command Palette.
-        */}
-      </footer>
-    </aside>
-  );
-}
-
-/**
- * PR-EMPTY-STATE-COMPONENT-0: shared empty-state container. Folds the
- * 4 visual duplicates (skills empty / sessions empty / module fallbacks /
- * plan reminders empty) into a single declaration so the next empty
- * surface lands consistent by default and the icon-sizing /
- * paragraph-spacing / CTA-placement decisions only live in one
- * place. The `.maka-empty-state*` CSS family is unchanged.
- *
- * Body accepts `ReactNode` so callers can keep inline `<code>` for
- * the skills install instructions; CTAs are rendered as the canonical
- * `.maka-button.maka-empty-state-cta` so we never grow a competing
- * pile of "empty-state action variants".
- */
-interface EmptyStateProps {
-  Icon: typeof Search;
-  title: string;
-  body: ReactNode;
-  cta?: { label: string; onClick: () => void; disabled?: boolean };
-  secondaryCta?: { label: string; onClick: () => void; disabled?: boolean };
-  /** Optional extra class on the container (e.g. `maka-plan-empty`). */
-  extraClassName?: string;
-  /** Optional `data-empty-view` passthrough for visual-smoke selectors. */
-  dataEmptyView?: string;
-}
-
-function EmptyState(props: EmptyStateProps) {
-  const className = cn(
-    'maka-empty-state rounded-xl border-border bg-card/70 p-8 text-card-foreground shadow-maka-panel',
-    props.extraClassName,
-  );
-  return (
-    <Empty className={className} data-empty-view={props.dataEmptyView}>
-      <EmptyHeader>
-        <EmptyMedia variant="icon" className="maka-empty-state-media">
-          <props.Icon className="maka-empty-state-icon size-6 text-muted-foreground" strokeWidth={1.5} />
-        </EmptyMedia>
-        <EmptyTitle className="maka-empty-state-title">{props.title}</EmptyTitle>
-        <EmptyDescription className="maka-empty-state-body">{props.body}</EmptyDescription>
-      </EmptyHeader>
-      {(props.cta || props.secondaryCta) && (
-        <EmptyContent className="maka-empty-state-actions mt-0">
-          {props.cta && (
-            <UiButton
-              className="maka-button maka-empty-state-cta"
-              type="button"
-              onClick={props.cta.onClick}
-              disabled={props.cta.disabled}
-            >
-              {props.cta.label}
-            </UiButton>
-          )}
-          {props.secondaryCta && (
-            <UiButton
-              variant="ghost"
-              className="maka-button maka-empty-state-cta"
-              type="button"
-              onClick={props.secondaryCta.onClick}
-              disabled={props.secondaryCta.disabled}
-            >
-              {props.secondaryCta.label}
-            </UiButton>
-          )}
-        </EmptyContent>
-      )}
-    </Empty>
-  );
-}
 
 function SkillLibraryPanel(props: {
   skills?: SkillEntry[];
@@ -1187,52 +588,7 @@ function formatSkillLibraryDescription(skill: SkillEntry): string | undefined {
   return '打开技能文件查看适用场景。';
 }
 
-export function CapabilityAuditStrip(props: { report: CapabilityAuditReport; focus: 'skills' | 'automations' }) {
-  const report = props.report;
-  const sourceCopy = report.summary.sourceCount === 0
-    ? '0 个来源'
-    : `${report.summary.readySourceCount}/${report.summary.sourceCount} 来源就绪`;
-  const skillCopy = `${report.summary.enabledSkillCount}/${report.summary.skillCount} 技能启用`;
-  const automationCopy = `${report.summary.enabledAutomationCount}/${report.summary.automationCount} 自动化启用`;
-  const riskCopy = capabilityAuditRiskCopy(report);
-  const primaryCopy = props.focus === 'skills'
-    ? `${report.summary.declaredToolKindCount} 类声明工具`
-    : `${report.summary.executableAutomationCount} 个可执行自动化`;
 
-  return (
-    <section className="maka-capability-audit-strip" aria-label="能力审计摘要">
-      <div className="maka-capability-audit-copy">
-        <span className="maka-capability-audit-kicker">能力审计</span>
-        <strong>{primaryCopy}</strong>
-        <small>{riskCopy}</small>
-      </div>
-      <dl className="maka-capability-audit-metrics" aria-label="来源、技能、自动化状态">
-        <div>
-          <dt>来源</dt>
-          <dd>{sourceCopy}</dd>
-        </div>
-        <div>
-          <dt>技能</dt>
-          <dd>{skillCopy}</dd>
-        </div>
-        <div>
-          <dt>自动化</dt>
-          <dd>{automationCopy}</dd>
-        </div>
-      </dl>
-    </section>
-  );
-}
-
-function capabilityAuditRiskCopy(report: CapabilityAuditReport): string {
-  const issues: string[] = [];
-  if (report.summary.needsAuthSourceCount > 0) issues.push(`${report.summary.needsAuthSourceCount} 个来源等待授权`);
-  if (report.summary.errorSourceCount > 0) issues.push(`${report.summary.errorSourceCount} 个来源异常`);
-  if (report.summary.failedAutomationCount > 0) issues.push(`${report.summary.failedAutomationCount} 个自动化上次失败`);
-  if (report.summary.skippedAutomationCount > 0) issues.push(`${report.summary.skippedAutomationCount} 个自动化上次跳过`);
-  if (issues.length === 0) return 'Skill 声明工具只作为权限请求展示；不会放大当前会话权限。';
-  return issues.join(' · ');
-}
 
 function SkillsModuleMain(props: {
   skills?: SkillEntry[];
@@ -1339,43 +695,7 @@ function SkillsModuleMain(props: {
   );
 }
 
-/**
- * PR-DAILY-REVIEW-MVP-0: bridge handed in by `main.tsx`. Keeps
- * `@maka/ui` out of `window.maka` — the renderer wires
- * `(offsetDays) => window.maka.dailyReview.day(offsetDays)` and the
- * UI layer is reusable in fixtures / visual smoke / future surfaces
- * (e.g. a desktop notification renderer).
- */
-interface DailyReviewBridge {
-  fetchDay(offsetDays: number, daySpan?: number): Promise<DailyReviewSummary>;
-  /**
-   * PR-DAILY-REVIEW-FULL-0 — optional pipeline methods. Renderer checks
-   * for presence before exposing the matching UI. When undefined, the
-   * panel still works as the MVP telemetry view.
-   */
-  runOnce?(opts: { mode: DailyReviewMode; modelKey?: string }): Promise<{ archiveId: string }>;
-  modelOptions?: ReadonlyArray<SettingsSelectOption<string>>;
-  listArchives?(): Promise<DailyReviewArchiveSummary[]>;
-  getArchive?(archiveId: string): Promise<DailyReviewArchive>;
-  deleteArchive?(archiveId: string): Promise<void>;
-  fetchConfig?(): Promise<DailyReviewConfig>;
-  updateConfig?(patch: Partial<DailyReviewConfig>): Promise<DailyReviewConfig>;
-}
 
-/**
- * Local-only daily summary view. Renders today by default; the
- * left/right arrows step through `offsetDays`. No LLM call — the
- * bullet list of sessions / top tools / top models is the whole
- * value-prop. Future PR can layer a generated narrative on top.
- *
- * borrow: external "today" digest concept (read-only summary).
- * diverge: no cron, no auto-push, no memory promotion (privacy default).
- */
-type DailyReviewMarkdownActionInput = {
-  markdown: string;
-  label: string;
-  summary: DailyReviewSummary;
-};
 type DailyReviewArchiveSectionKey = keyof DailyReviewArchive['sections'];
 
 const DAILY_REVIEW_ARCHIVE_SECTION_LABEL: Record<DailyReviewArchiveSectionKey, string> = {
@@ -2792,655 +2112,14 @@ function PlanReminderPanel(props: {
   );
 }
 
-/**
- * PR-SIDEBAR-IA-0 Phase 2 fixup (xuan `91401163` + kenji `6465cf22`,
- * `7c320898`) + Phase 3 P0 fixup (WAWQAQ msg `d53852ac`, xuan
- * `558f1356`, kenji `3ddc91fe`): Search modal SHELL.
- *
- * Renders the real thread-search dialog: local query state,
- * debounced `search:thread` IPC, result list, incognito/error states,
- * and shell-owned navigation. It never writes history and never
- * constructs `maka://session` URIs.
- *
- * Lifecycle contract: SearchModal MUST be conditionally mounted by
- * the parent (`{open && <SearchModal onClose={...} />}`), NOT
- * always-mounted with an `open` prop. The previous pattern
- * (`<SearchModal open=... />` with an internal `if (!open) return
- * null`) sat hooks before a conditional return; while React allows
- * this in principle, in production WAWQAQ hit a React #310 hook
- * order mismatch via the same surface (msg `d53852ac`). Matching
- * `KeyboardHelpModal`'s conditional-mount pattern eliminates the
- * "hooks before early return" class of bug entirely — there's no
- * way for a future hook addition to drift past a stale return
- * statement.
- *
- * Gate per kenji `7c320898`:
- *   - role="dialog" / aria-modal="true" / explicit title.
- *   - Esc and close button close the modal.
- *   - Focus enters the modal on open; returns to the trigger on close.
- *   - Modal calls injected `searchThread` only; it does NOT store
- *     the query, write history, or route via internal URI strings.
- */
-/**
- * Dependency-injected search interface. Production wiring binds this
- * to `window.maka.search.thread`; tests pass an in-memory fake.
- *
- * The return type matches the IPC envelope exactly: either an array
- * of `SearchResult` (success path) or a `{ ok: false, reason, message }`
- * error envelope. Renderer never throws across the IPC boundary —
- * fail-closed paths return the error envelope and the modal renders
- * them as user-facing copy.
- */
-interface SearchModalDeps {
-  searchThread(request: SearchRequest): Promise<
-    | SearchResult[]
-    | { ok: false; reason: SearchErrorReason; message: string }
-  >;
-}
 
-function searchModalThrownErrorMessage(error: unknown): string {
-  return generalizedErrorMessageChinese(error, '搜索服务需要刷新，请重试。');
-}
 
-interface SearchModalCloseOptions {
-  restoreFocus?: boolean;
-}
 
-export function SearchModal(props: {
-  onClose(options?: SearchModalCloseOptions): void;
-  /**
-   * Navigate to a session (optionally scrolling to a specific turn).
-   * Provided by the application shell so the modal stays portable —
-   * navigation lives in the shell, not in @maka/ui.
-   *
-   * Per kenji `2844f64f` SEARCH gate: navigation MUST NOT construct
-   * `maka://session/<id>` URIs. The callback receives raw ids; the
-   * shell handles routing via existing session-pane state.
-   */
-  onNavigateToSession?(sessionId: string, turnId?: string): void;
-  /**
-   * Injected `search:thread` IPC. Production binds to
-   * `window.maka.search.thread`; tests supply a fake.
-   *
-   * Optional so the modal renders a degraded "search unavailable"
-   * state when the renderer cannot bind to the IPC (legacy / smoke
-   * fixture / preload not loaded). Without an injected deps the
-   * modal does NOT crash.
-   */
-  deps?: SearchModalDeps;
-}) {
-  const dialogRef = useRef<HTMLDivElement>(null);
 
-  // PR-UX-POLISH-1 commit 5 (kenji `2844f64f` SEARCH gate):
-  //   - `query` is local state ONLY (no localStorage / no IPC echo).
-  //   - `results` is the most recent successful response; older
-  //     responses are discarded by the inflight ticket guard so the
-  //     UI never shows stale data behind a newer query.
-  //   - `error` carries the IPC error envelope when present. We do
-  //     NOT raise it as a JS throw — the modal renders the message
-  //     copy and the gate's `incognito_active` / `invalid_query`
-  //     reasons trigger specific UI states (privacy banner / empty).
-  //   - `pending` reflects whether ANY IPC call is in flight. We do
-  //     NOT show a spinner if the query is empty (avoids flashing
-  //     loading state during typing).
-  const [query, setQuery] = useState('');
-  const [results, setResults] = useState<SearchResult[]>([]);
-  const [error, setError] = useState<{ reason: SearchErrorReason; message: string } | null>(null);
-  const [pending, setPending] = useState(false);
-  const [activeResultIndex, setActiveResultIndex] = useState(-1);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const resultRefs = useRef<Array<HTMLButtonElement | null>>([]);
-  const ticketRef = useRef(0);
-  const searchMountedRef = useRef(true);
-  const keyboardSelectionHandledRef = useRef(false);
-  const searchThread = props.deps?.searchThread;
-  useModalA11y(dialogRef, props.onClose, inputRef);
 
-  useEffect(() => {
-    searchMountedRef.current = true;
-    return () => {
-      searchMountedRef.current = false;
-      ticketRef.current += 1;
-    };
-  }, []);
 
-  // Debounced search: ~180ms after the user stops typing, send the
-  // request. Empty query clears state without an IPC roundtrip.
-  useEffect(() => {
-    if (!searchThread) return;
-    const trimmed = query.trim();
-    if (trimmed.length === 0) {
-      ticketRef.current += 1;
-      setResults([]);
-      setError(null);
-      setPending(false);
-      setActiveResultIndex(-1);
-      return;
-    }
-    const ticket = ++ticketRef.current;
-    setPending(true);
-    const handle = window.setTimeout(async () => {
-      try {
-        const response = await searchThread({
-          source: 'thread',
-          query: trimmed,
-          limit: 10,
-        });
-        if (!searchMountedRef.current) return;
-        if (ticket !== ticketRef.current) return; // newer query in flight
-        if (Array.isArray(response)) {
-          setResults(response);
-          setError(null);
-          setActiveResultIndex(-1);
-        } else {
-          setResults([]);
-          setError({ reason: response.reason, message: response.message });
-          setActiveResultIndex(-1);
-        }
-      } catch (err) {
-        if (!searchMountedRef.current) return;
-        if (ticket !== ticketRef.current) return;
-        // IPC layer should never throw, but defend anyway. Render as a
-        // generic provider_error so the user sees a coherent state.
-        setResults([]);
-        setError({
-          reason: 'provider_error',
-          message: searchModalThrownErrorMessage(err),
-        });
-        setActiveResultIndex(-1);
-      } finally {
-        if (searchMountedRef.current && ticket === ticketRef.current) setPending(false);
-      }
-    }, 180);
-    return () => window.clearTimeout(handle);
-  }, [query, searchThread]);
 
-  useEffect(() => {
-    if (activeResultIndex < 0) return;
-    resultRefs.current[activeResultIndex]?.scrollIntoView({ block: 'nearest' });
-  }, [activeResultIndex]);
 
-  function selectResult(result: SearchResult) {
-    if (!props.onNavigateToSession) return;
-    if (result.target?.kind !== 'thread') return;
-    props.onNavigateToSession(result.target.sessionId, result.target.turnId);
-    props.onClose({ restoreFocus: false });
-  }
-
-  function selectKeyboardResult() {
-    if (!showResults) return;
-    selectResult(results[activeResultIndex >= 0 ? activeResultIndex : 0]!);
-  }
-
-  function clearSearchState() {
-    ticketRef.current += 1;
-    setResults([]);
-    setError(null);
-    setPending(false);
-    setActiveResultIndex(-1);
-  }
-
-  function updateSearchQuery(nextQuery: string) {
-    setQuery(nextQuery);
-    if (nextQuery.trim().length === 0) {
-      clearSearchState();
-    }
-  }
-
-  function clearSearchQuery() {
-    setQuery('');
-    clearSearchState();
-    inputRef.current?.focus();
-  }
-
-  function focusSearchResult(index: number) {
-    window.requestAnimationFrame(() => {
-      resultRefs.current[index]?.focus({ preventScroll: true });
-    });
-  }
-
-  function moveActiveResult(delta: 1 | -1, options?: { focusResult?: boolean }) {
-    if (results.length === 0) return;
-    const next = activeResultIndex < 0
-      ? (delta > 0 ? 0 : results.length - 1)
-      : (activeResultIndex + delta + results.length) % results.length;
-    setActiveResultIndex(next);
-    if (options?.focusResult) focusSearchResult(next);
-  }
-
-  function jumpActiveResult(index: number, options?: { focusResult?: boolean }) {
-    if (results.length === 0) return;
-    const next = Math.max(0, Math.min(results.length - 1, index));
-    setActiveResultIndex(next);
-    if (options?.focusResult) focusSearchResult(next);
-  }
-
-  function keyboardKey(event: KeyboardEvent, keys: string[]) {
-    return keys.includes(event.key) || keys.includes(event.code);
-  }
-
-  function handleResultKeyDown(event: KeyboardEvent<HTMLButtonElement>, index: number, result: SearchResult) {
-    if (keyboardKey(event, ['Enter', 'Return', 'Space', ' '])) {
-      event.preventDefault();
-      selectResult(result);
-      return;
-    }
-    if (keyboardKey(event, ['ArrowDown', 'Down'])) {
-      event.preventDefault();
-      moveActiveResult(1, { focusResult: true });
-      return;
-    }
-    if (keyboardKey(event, ['ArrowUp', 'Up'])) {
-      event.preventDefault();
-      moveActiveResult(-1, { focusResult: true });
-      return;
-    }
-    if (keyboardKey(event, ['Home'])) {
-      event.preventDefault();
-      jumpActiveResult(0, { focusResult: true });
-      return;
-    }
-    if (keyboardKey(event, ['End'])) {
-      event.preventDefault();
-      jumpActiveResult(results.length - 1, { focusResult: true });
-      return;
-    }
-    if (keyboardKey(event, ['Escape'])) {
-      event.preventDefault();
-      props.onClose();
-      return;
-    }
-    if (index !== activeResultIndex) {
-      setActiveResultIndex(index);
-    }
-  }
-
-  const incognitoBlocked = error?.reason === 'incognito_active';
-  const trimmed = query.trim();
-  const showResults = !error && trimmed.length > 0 && !pending && results.length > 0;
-  const showEmpty = !error && trimmed.length > 0 && !pending && results.length === 0;
-  const activeResultId = showResults && activeResultIndex >= 0 ? `maka-search-modal-result-${activeResultIndex}` : undefined;
-  const resultsTruncated = showResults && results.some((result) => result.truncated === true);
-
-  return (
-    <DialogRoot
-      open
-      onOpenChange={(open) => {
-        if (!open) props.onClose();
-      }}
-    >
-      <DialogContent
-        ref={dialogRef}
-        className="maka-modal maka-search-modal w-[min(92vw,640px)] p-0"
-        aria-labelledby="maka-search-modal-title"
-        showClose={false}
-      >
-        <header className="maka-search-modal-header">
-          <h2 id="maka-search-modal-title" className="maka-search-modal-title">搜索</h2>
-          <DialogClose
-            render={<UiButton variant="quiet" size="icon-sm" />}
-            type="button"
-            className="maka-search-modal-close"
-            onClick={() => props.onClose()}
-            aria-label="关闭搜索"
-          >
-            <X size={16} strokeWidth={1.8} aria-hidden="true" />
-          </DialogClose>
-        </header>
-        <InputGroup className="maka-search-modal-input-row" aria-label="搜索会话">
-          <InputGroupAddon>
-            <Search size={16} strokeWidth={1.75} aria-hidden="true" className="maka-search-modal-input-icon" />
-          </InputGroupAddon>
-          <InputGroupInput
-            ref={inputRef}
-            type="search"
-            className="maka-search-modal-input"
-            placeholder="搜索会话标题和内容…"
-            aria-label="搜索会话标题和内容"
-            aria-controls={showResults ? 'maka-search-modal-results' : undefined}
-            aria-activedescendant={activeResultId}
-            value={query}
-            onChange={(event) => updateSearchQuery(event.currentTarget.value)}
-            onKeyDown={(event) => {
-              if (keyboardKey(event, ['Escape']) && query) {
-                event.preventDefault();
-                clearSearchQuery();
-                return;
-              }
-              if (keyboardKey(event, ['ArrowDown', 'Down']) && showResults) {
-                event.preventDefault();
-                moveActiveResult(1, { focusResult: true });
-                return;
-              }
-              if (keyboardKey(event, ['ArrowUp', 'Up']) && showResults) {
-                event.preventDefault();
-                moveActiveResult(-1, { focusResult: true });
-                return;
-              }
-              if (keyboardKey(event, ['Home']) && showResults) {
-                event.preventDefault();
-                jumpActiveResult(0, { focusResult: true });
-                return;
-              }
-              if (keyboardKey(event, ['End']) && showResults) {
-                event.preventDefault();
-                jumpActiveResult(results.length - 1, { focusResult: true });
-                return;
-              }
-              if (keyboardKey(event, ['Enter', 'Return']) && showResults) {
-                event.preventDefault();
-                keyboardSelectionHandledRef.current = true;
-                selectKeyboardResult();
-              }
-            }}
-            onKeyUp={(event) => {
-              if (keyboardKey(event, ['Enter', 'Return']) && keyboardSelectionHandledRef.current) {
-                if (showResults) event.preventDefault();
-                keyboardSelectionHandledRef.current = false;
-                return;
-              }
-              if (keyboardKey(event, ['Enter', 'Return']) && showResults) {
-                event.preventDefault();
-                selectKeyboardResult();
-              }
-            }}
-            autoComplete="off"
-            spellCheck={false}
-          />
-          {query.length > 0 && (
-            <InputGroupAddon align="inline-end">
-              <UiButton
-                variant="quiet"
-                size="icon-sm"
-                type="button"
-                className="maka-search-modal-clear"
-                aria-label="清空搜索"
-                onClick={clearSearchQuery}
-              >
-                <X size={14} strokeWidth={1.8} aria-hidden="true" />
-              </UiButton>
-            </InputGroupAddon>
-          )}
-        </InputGroup>
-        <div className="maka-search-modal-body" role="region" aria-label="搜索状态和结果" aria-live="polite">
-          {!searchThread && (
-            <p className="maka-search-modal-placeholder">
-              当前环境无法连接搜索后端，请稍后重试。
-            </p>
-          )}
-          {searchThread && incognitoBlocked && (
-            <div className="maka-search-modal-state" data-tone="info">
-              <p>隐私模式已关闭搜索。</p>
-              <p className="maka-search-modal-state-detail">
-                关闭隐私模式后可以继续按关键词查找历史对话。
-              </p>
-            </div>
-          )}
-          {searchThread && !incognitoBlocked && error && (
-            <div className="maka-search-modal-state" data-tone="warning">
-              <p>搜索暂时无法完成。</p>
-              <p className="maka-search-modal-state-detail">{error.message}</p>
-            </div>
-          )}
-          {searchThread && !error && trimmed.length === 0 && (
-            <p className="maka-search-modal-placeholder">
-              开始输入以按关键词查找历史对话。结果只包含会话标题和内容文本，不进入网络。
-            </p>
-          )}
-          {searchThread && pending && trimmed.length > 0 && (
-            <p className="maka-search-modal-placeholder" aria-live="polite">
-              正在搜索…
-            </p>
-          )}
-          {showEmpty && (
-            <p className="maka-search-modal-placeholder">
-              没有匹配的会话标题或内容。换个关键词试试。
-            </p>
-          )}
-          {showResults && (
-            <>
-              <div className="maka-search-modal-result-summary" aria-live="polite">
-                <span>找到 {results.length} 条匹配</span>
-                {resultsTruncated && <span>结果较多，已显示前 {results.length} 条</span>}
-              </div>
-              <ul id="maka-search-modal-results" className="maka-search-modal-results" role="listbox" aria-label="搜索结果">
-                {results.map((result, index) => (
-                  <li key={`${result.target?.kind === 'thread' ? result.target.sessionId : index}-${index}`}>
-                    <UiButton
-                      variant="ghost"
-                      ref={(node) => { resultRefs.current[index] = node as HTMLButtonElement | null; }}
-                      id={`maka-search-modal-result-${index}`}
-                      type="button"
-                      role="option"
-                      aria-selected={activeResultIndex === index}
-                      tabIndex={-1}
-                      className="maka-search-modal-result"
-                      data-active={activeResultIndex === index ? 'true' : undefined}
-                      onClick={() => selectResult(result)}
-                      onKeyDown={(event) => handleResultKeyDown(event, index, result)}
-                      onFocus={() => setActiveResultIndex(index)}
-                      onMouseEnter={() => setActiveResultIndex(index)}
-                      disabled={!props.onNavigateToSession || result.target?.kind !== 'thread'}
-                    >
-                      <div className="maka-search-modal-result-title">{result.title}</div>
-                      {result.summary && <div className="maka-search-modal-result-meta">{result.summary}</div>}
-                      {result.snippet && (
-                        // Plain text only — IPC already redacts secrets
-                        // and the snippet is bounded by SNIPPET_MAX_CODE_POINTS.
-                        // No markdown rendering, no <img>, no <a href> —
-                        // per kenji SEARCH gate (no path / no URL exposure).
-                        <div className="maka-search-modal-result-snippet">{renderSearchSnippet(result.snippet, trimmed)}</div>
-                      )}
-                    </UiButton>
-                  </li>
-                ))}
-              </ul>
-            </>
-          )}
-        </div>
-      </DialogContent>
-    </DialogRoot>
-  );
-}
-
-function renderSearchSnippet(snippet: string, query: string): ReactNode {
-  const needle = query.trim();
-  if (!needle) return snippet;
-  const haystack = snippet.toLocaleLowerCase();
-  const lowerNeedle = needle.toLocaleLowerCase();
-  const parts: ReactNode[] = [];
-  let cursor = 0;
-  let matchIndex = haystack.indexOf(lowerNeedle);
-  while (matchIndex !== -1) {
-    if (matchIndex > cursor) {
-      parts.push(snippet.slice(cursor, matchIndex));
-    }
-    const end = matchIndex + needle.length;
-    parts.push(
-      <mark key={`${matchIndex}-${end}`} className="maka-search-modal-snippet-hit">
-        {snippet.slice(matchIndex, end)}
-      </mark>,
-    );
-    cursor = end;
-    matchIndex = haystack.indexOf(lowerNeedle, cursor);
-  }
-  if (cursor < snippet.length) parts.push(snippet.slice(cursor));
-  return parts.length > 0 ? parts : snippet;
-}
-
-/**
- * Render an ordered list of session groups, supporting collapsibility
- * per group. Used by SessionListPanel for both the legacy date-bucket
- * grouping and the new status-driven grouping (PR109b).
- *
- * Each group has a header row with the group label + count. Collapsible
- * groups show a chevron and toggle expanded state via local state.
- * Expanded state is keyed on group `key` so the same group keeps its
- * state across re-renders (e.g., archived stays collapsed even when
- * sidebar refreshes).
- */
-function SessionListGroups(props: {
-  groups: ReadonlyArray<{
-    key: string;
-    label: string;
-    sessions: SessionSummary[];
-    collapsible: boolean;
-    defaultExpanded: boolean;
-  }>;
-  activeId?: string;
-  streamingSessionIds?: Set<string>;
-  staleSessionIds?: Set<string>;
-  onSelectSession(sessionId: string): void;
-  rowActions?: SessionRowActions;
-}) {
-  const [expandedByKey, setExpandedByKey] = useState<Record<string, boolean>>(() => {
-    const out: Record<string, boolean> = {};
-    for (const g of props.groups) out[g.key] = g.defaultExpanded;
-    return out;
-  });
-  // Ensure newly-appearing groups inherit their defaultExpanded value
-  // without overriding user-toggled state.
-  useEffect(() => {
-    setExpandedByKey((current) => {
-      const next = { ...current };
-      let changed = false;
-      for (const g of props.groups) {
-        if (!(g.key in next)) {
-          next[g.key] = g.defaultExpanded;
-          changed = true;
-        }
-      }
-      return changed ? next : current;
-    });
-  }, [props.groups]);
-  return (
-    <>
-      {props.groups.map((group) => {
-        const expanded = expandedByKey[group.key] ?? group.defaultExpanded;
-        const toggle = () =>
-          setExpandedByKey((current) => ({ ...current, [group.key]: !expanded }));
-        return (
-          <div key={group.key} className="maka-list-group" data-collapsible={group.collapsible || undefined}>
-            {group.collapsible ? (
-              /* PR-LIST-GROUP-TOGGLE-PRIMITIVE-0 (round 10/30):
-                 disclosure-pattern toggle (aria-expanded +
-                 aria-controls). Routed through UiButton so the
-                 collapsible group header shares the same
-                 focus-visible + `:active` contract as every
-                 other interactive surface in the session list. */
-              <UiButton
-                type="button"
-                variant="quiet"
-                className="maka-list-group-label maka-list-group-toggle"
-                onClick={toggle}
-                aria-expanded={expanded}
-                aria-controls={`maka-list-group-body-${group.key}`}
-              >
-                <ChevronRight
-                  size={12}
-                  strokeWidth={2}
-                  aria-hidden="true"
-                  style={{
-                    transform: expanded ? 'rotate(90deg)' : undefined,
-                    transition: 'transform 140ms var(--ease-out-strong)',
-                  }}
-                />
-                <span>{group.label}</span>
-                {/* Collapsed history buckets keep a subdued count so users
-                  can tell whether expanding the group is worth it. Open
-                  groups intentionally omit counts to keep the rail flat. */}
-                <span className="maka-list-group-count">（{group.sessions.length}）</span>
-              </UiButton>
-            ) : (
-              <div className="maka-list-group-label">
-                <span>{group.label}</span>
-              </div>
-            )}
-            {expanded && (
-              <div id={`maka-list-group-body-${group.key}`}>
-                {group.sessions.map((session) => (
-                  <SessionRow
-                    key={session.id}
-                    session={session}
-                    active={session.id === props.activeId}
-                    streaming={props.streamingSessionIds?.has(session.id) ?? false}
-                    stale={props.staleSessionIds?.has(session.id) ?? false}
-                    onSelect={props.onSelectSession}
-                    actions={props.rowActions}
-                  />
-                ))}
-              </div>
-            )}
-          </div>
-        );
-      })}
-    </>
-  );
-}
-
-/**
- * Small inline icon next to the session name representing its
- * lifecycle status (PR109b, design-system §9.8). Hidden for `active`
- * since that's the default and would add visual noise to most rows.
- *
- * `aborted` is rendered as muted history: not an error, not active,
- * and not silently swallowed.
- *
- * Caller is expected to pass a session with a SessionStatus from
- * `@maka/core` — typed as the SessionSummary from props avoids
- * pulling the core type into this file's import list.
- */
-function SessionStatusIcon(props: { session: SessionSummary }) {
-  const { session } = props;
-  const status = session.status;
-  // Active is the default; no icon to reduce noise. Aborted retains a
-  // muted icon (per @kenji review on PR109b — aborted is dormant
-  // history that must remain visible, not silently swallowed as active).
-  if (status === 'active') return null;
-  const Icon = STATUS_ICON_BY_STATUS[status as keyof typeof STATUS_ICON_BY_STATUS];
-  if (!Icon) return null;
-  const label = STATUS_LABEL_BY_STATUS[status as keyof typeof STATUS_LABEL_BY_STATUS];
-  const tone = STATUS_TONE_BY_STATUS[status as keyof typeof STATUS_TONE_BY_STATUS];
-  // `blocked` may attach a reason; we surface the generalized text in
-  // the tooltip without exposing the raw enum identifier (per @kenji
-  // i18n contract). The reason mapping lives in the renderer side; this
-  // file knows only the status itself, so the tooltip is just the
-  // status label.
-  const blockedDetail = status === 'blocked' && session.blockedReason
-    ? BLOCKED_REASON_TOOLTIP[session.blockedReason as keyof typeof BLOCKED_REASON_TOOLTIP] ?? null
-    : null;
-  const title = blockedDetail ? `${label} · ${blockedDetail}` : label;
-  return (
-    <span
-      className="maka-list-row-status-icon"
-      data-tone={tone}
-      data-status={status}
-      aria-label={title}
-      title={title}
-    >
-      <Icon size={12} strokeWidth={2} aria-hidden="true" />
-    </span>
-  );
-}
-
-/**
- * PawWork-style sidebar attention priority: asking/busy/error outrank unread,
- * and unread outranks plain time. The status icon beside the name already
- * carries asking/busy/error in Maka, so the right slot only shows the unread
- * dot when no higher-priority row state is active.
- */
-function shouldShowSessionUnreadDot(session: SessionSummary, streaming: boolean, active: boolean): boolean {
-  if (active) return false;
-  if (!session.hasUnread) return false;
-  if (streaming) return false;
-  return !SIDEBAR_UNREAD_SUPPRESSED_STATUSES.has(session.status);
-}
-
-const SIDEBAR_UNREAD_SUPPRESSED_STATUSES = new Set<string>([
-  'running',
-  'waiting_for_user',
-  'blocked',
-]);
 
 /**
  * Lifecycle status badge in the chat header (PR109b §9.8). Visual
@@ -3469,398 +2148,15 @@ function SessionStatusBadge(props: {
   );
 }
 
-// Keep these maps in sync with `apps/desktop/src/renderer/session-status-presentation.ts`.
-// The presentation helper is the authoritative source; we duplicate the
-// minimum subset here to keep @maka/ui independent of the renderer
-// workspace.
-const STATUS_ICON_BY_STATUS = {
-  running: Loader2,
-  waiting_for_user: Hourglass,
-  blocked: ShieldAlert,
-  review: Eye,
-  done: CircleCheckBig,
-  archived: Archive,
-  aborted: Ban,
-} as const;
 
-const STATUS_LABEL_BY_STATUS = {
-  running: '进行中',
-  waiting_for_user: '等你确认',
-  blocked: '已阻塞',
-  review: '待审核',
-  done: '已完成',
-  archived: '已归档',
-  aborted: '已中止',
-} as const;
 
-// `blocked` was 'destructive' (red), which read as a hard error in the
-// chat header even when the session was just waiting on permission or a
-// connection retry. The chat top-right cluster sits visually alongside
-// monochrome quiet-icon buttons, so the bright red pill clashed. Most
-// blocked sessions are recoverable (permission_required, auth retry,
-// missing connection), so 'warning' (warm yellow) is the honest tone —
-// "you need to do something" rather than "this failed". The destructive
-// red is reserved for hard failures (e.g. permanent connection / auth
-// rejection), which surface through ChatHeaderAlertBadge instead.
-const STATUS_TONE_BY_STATUS = {
-  running: 'accent',
-  waiting_for_user: 'warning',
-  blocked: 'warning',
-  review: 'info',
-  done: 'success',
-  archived: 'muted',
-  aborted: 'muted',
-} as const;
 
-const BLOCKED_REASON_TOOLTIP = {
-  NO_REAL_CONNECTION: '等待配置可用模型连接',
-  auth: '需要重新登录',
-  permission_required: '等待权限确认',
-  tool_failed: '工具调用失败',
-  unknown: '运行中断，可重试',
-} as const;
 
 const SCROLL_BOTTOM_THRESHOLD = 64; // px
 
 // Back-compat alias for the helper introduced in PR-UI-14.
 const detectPromptSuggestionLocale = detectUiLocale;
 
-function SessionRow(props: {
-  session: SessionSummary;
-  active: boolean;
-  /** This session has a live streaming delta in flight. */
-  streaming?: boolean;
-  /**
-   * This session's backend / connection is stale (FakeBackend or a removed
-   * connection slug). Dims the row + renders a small "已过期" pill so the
-   * user can spot broken sessions in the list before clicking in.
-   */
-  stale?: boolean;
-  onSelect(sessionId: string): void;
-  actions?: SessionRowActions;
-}) {
-  const { session, active, streaming, stale, actions, onSelect } = props;
-  const [editing, setEditing] = useState(false);
-  const [actionsVisible, setActionsVisible] = useState(false);
-  const [pendingAction, setPendingAction] = useState<SessionRowActionId | null>(null);
-  const rowMountedRef = useRef(true);
-  const pendingActionRef = useRef<SessionRowActionId | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  // PR-FE-BUG-HUNT-11: Escape on the rename input has to suppress the
-  // blur-fires-on-unmount commit. Without this ref, the sequence
-  //   type → Escape → setEditing(false) → input unmounts → blur fires
-  //   with the typed value → commitRename(typed) → rename happens
-  // would silently commit the user's typed value despite the cancel.
-  const escapeCancelledRef = useRef(false);
-  const actionBusy = pendingAction !== null;
-  const actionTabIndex = actionsVisible ? 0 : -1;
-
-  useEffect(() => {
-    rowMountedRef.current = true;
-    return () => {
-      rowMountedRef.current = false;
-      pendingActionRef.current = null;
-    };
-  }, []);
-
-  // Auto-focus + select-all when the row enters edit mode so the user can
-  // overwrite the current name without an extra Cmd+A.
-  useEffect(() => {
-    if (!editing) return;
-    const input = inputRef.current;
-    if (!input) return;
-    input.focus();
-    input.select();
-  }, [editing]);
-
-  const stopPropagation = (event: MouseEvent<HTMLButtonElement>) => {
-    event.stopPropagation();
-  };
-
-  function startRename(event: MouseEvent<HTMLButtonElement>) {
-    stopPropagation(event);
-    if (!actions || pendingActionRef.current) return;
-    setEditing(true);
-  }
-
-  function runRowAction(actionId: SessionRowActionId, action: () => void | Promise<void>) {
-    if (pendingActionRef.current) return;
-    pendingActionRef.current = actionId;
-    setPendingAction(actionId);
-    void Promise.resolve().then(action).finally(() => {
-      pendingActionRef.current = null;
-      if (rowMountedRef.current) setPendingAction(null);
-    });
-  }
-
-  function commitRename(rawValue: string) {
-    const trimmed = rawValue.trim();
-    setEditing(false);
-    if (!trimmed || trimmed === session.name) return;
-    if (!actions) return;
-    runRowAction('rename', () => actions.onRename(session.id, trimmed));
-  }
-
-  function handleDelete(event: MouseEvent<HTMLButtonElement>) {
-    stopPropagation(event);
-    if (!actions) return;
-    // Delegation: the App-level handler owns the confirmation flow via the
-    // toast system (PR24), so SessionRow stays presentation-only.
-    runRowAction('delete', () => actions.onDelete(session.id));
-  }
-
-  function handleRowBlur(event: FocusEvent<HTMLDivElement>) {
-    if (event.currentTarget.contains(event.relatedTarget as Node | null)) return;
-    setActionsVisible(false);
-  }
-
-  return (
-    <div
-      className="maka-list-row"
-      data-active={active}
-      data-editing={editing}
-      data-streaming={streaming ? 'true' : undefined}
-      data-stale={stale ? 'true' : undefined}
-      onMouseEnter={() => setActionsVisible(true)}
-      onMouseLeave={(event) => {
-        if (event.currentTarget.contains(document.activeElement)) return;
-        setActionsVisible(false);
-      }}
-      onFocus={() => setActionsVisible(true)}
-      onBlur={handleRowBlur}
-    >
-      {editing ? (
-        <form
-          className="maka-list-row-main"
-          onSubmit={(event) => {
-            event.preventDefault();
-            commitRename(inputRef.current?.value ?? '');
-          }}
-        >
-          <div>
-            <input
-              ref={inputRef}
-              className="maka-list-row-rename-input"
-              defaultValue={session.name}
-              maxLength={80}
-              aria-label="重命名对话"
-              onBlur={(event) => {
-                // PR-FE-BUG-HUNT-11: skip the commit when the blur was
-                // caused by Escape cancelling edit mode (input unmounts
-                // → blur fires with the typed value otherwise).
-                if (escapeCancelledRef.current) {
-                  escapeCancelledRef.current = false;
-                  return;
-                }
-                commitRename(event.currentTarget.value);
-              }}
-              onKeyDown={(event) => {
-                // IME guard so committing CJK characters with Enter doesn't
-                // submit the rename before the user is done.
-                if (event.nativeEvent.isComposing || event.key === 'Process') return;
-                if (event.key === 'Escape') {
-                  event.preventDefault();
-                  escapeCancelledRef.current = true;
-                  setEditing(false);
-                }
-              }}
-              autoComplete="off"
-              spellCheck={false}
-            />
-            <div className="maka-list-row-meta">{formatSessionMeta(session)}</div>
-          </div>
-        </form>
-      ) : (
-        // PR-SIDEBAR-IA-0 Phase 3 (WAWQAQ `14ed98b5` "list 很丑、很肥很
-        // 臃肿"; xuan `6b28984e` Phase 2 sign-off + Phase 3 32-40px
-        // target; xuan `2d4526b5` tightening: NO native title= snippet,
-        // title is ONLY for name truncation): slim row.
-        //
-        // The button is the row's hit target. The native `title=`
-        // attribute carries ONLY the session name so it serves as a
-        // truncation tooltip when the name overflows the row. The
-        // `lastMessagePreview` snippet is intentionally NOT exposed
-        // here — per xuan `2d4526b5`, snippet visibility is a
-        // separate, deliberate design (future PR), not a Phase 3
-        // afterthought via native tooltip.
-        //
-        // `data-active` on the row controls the active-state accent
-        // rail + bg tint via CSS; the row's `name` cluster also
-        // recolors to accent on selected so the row reads as
-        // "current" without a heavy full-bg pill.
-        /* PR-SESSION-ROW-MAIN-PRIMITIVE-0 (round 14/30): the
-           single most-clicked button in the app — every session
-           row's main click target. Routed through UiButton with
-           size="nav" (round-12 enabler) so the bespoke
-           `.maka-list-row-main` density (32-40px height, custom
-           padding, grid layout with text/meta columns) stays the
-           source of truth, while the primitive contributes the
-           shared `:active scale`, focus-visible, and
-           disabled-state contract. */
-        <UiButton
-          variant="quiet"
-          size="nav"
-          className="maka-list-row-main"
-          type="button"
-          data-session-id={session.id}
-          aria-current={active ? 'true' : undefined}
-          title={session.name}
-          onClick={() => onSelect(session.id)}
-          onDoubleClick={(event) => {
-            event.stopPropagation();
-            if (actions && !pendingActionRef.current) setEditing(true);
-          }}
-        >
-          {/*
-            PR-SIDEBAR-IA-0 Phase 3 layout (xuan `2d4526b5`):
-              [.maka-list-row-text  (col 1: minmax(0,1fr))] [meta/unread  (col 2: auto)]
-            The text container holds the name cluster (status icons +
-            name + stale pill) and truncates via min-width: 0. The
-            meta column sits at the inline-end with a clear gap so
-            "会话 02" doesn't run into "0m ago".
-          */}
-          <div className="maka-list-row-text">
-            <div className="maka-list-row-name">
-              {streaming && (
-                <span
-                  className="maka-list-row-streaming-dot"
-                  aria-label="正在响应"
-                  title="对话正在流式响应中"
-                />
-              )}
-              <SessionStatusIcon session={session} />
-              <span>{session.name}</span>
-              {stale && (
-                <span
-                  className="maka-list-row-stale-pill"
-                  // The pill semantics match the chat-header banner: the
-                  // session uses a backend / connection that no longer exists,
-                  // but @xuan's send-path silent rebind will swap to the
-                  // default on send. Tooltip explains why.
-                  title="此会话使用的模型连接已不可用，发送时会切换到默认连接"
-                  aria-label="会话已过期"
-                >
-                  已过期
-                </span>
-              )}
-            </div>
-          </div>
-          {/*
-            PR-SIDEBAR-IA-0 Phase 3 (xuan `2d4526b5`): snippet preview
-            (`.maka-list-row-preview`) is no longer rendered in the
-            default DOM AND is no longer exposed via native `title=`
-            tooltip. Snippet visibility is deliberately deferred to a
-            future PR with its own hover/focus detail design.
-            `formatSessionMeta` shows the relative time inline in the
-            row's `auto` grid column (sibling of `.maka-list-row-text`,
-            not nested inside it — required for proper gap + alignment).
-            The unread dot replaces the time only when no higher-priority
-            row state is active. Borrowed from PawWork's sidebar priority:
-            asking/busy/error outrank unread; unread outranks plain time.
-          */}
-          {shouldShowSessionUnreadDot(session, Boolean(streaming), active) ? (
-            <span className="maka-list-row-unread" aria-label="未读消息" />
-          ) : (
-            <span className="maka-list-row-meta">{formatSessionMeta(session)}</span>
-          )}
-        </UiButton>
-      )}
-      {actions && !editing && (
-        <div
-          className="maka-list-row-actions"
-          aria-label="对话操作"
-          aria-hidden={actionsVisible ? undefined : 'true'}
-          data-visible={actionsVisible ? 'true' : undefined}
-        >
-          {/* PR-SESSION-ROW-ACTIONS-PRIMITIVE-0 (round 8/30):
-              four hover-revealed action buttons routed through
-              UiButton variant="quiet" size="icon-sm". Custom
-              `.maka-list-row-action` still owns the overlay
-              positioning + reveal animation; primitive carries
-              the disabled, focus-visible, and `:active` contract.
-              The danger variant only adds a destructive color
-              tint via class override, not a different primitive
-              variant — keeps the overlay shape uniform. */}
-          <UiButton
-            type="button"
-            variant="quiet"
-            size="icon-sm"
-            className="maka-list-row-action"
-            tabIndex={actionTabIndex}
-            onClick={(event) => {
-              stopPropagation(event);
-              runRowAction('flag', () => actions.onToggleFlag(session.id, !session.isFlagged));
-            }}
-            aria-label={session.isFlagged ? '取消置顶对话' : '置顶对话'}
-            aria-busy={pendingAction === 'flag' ? 'true' : undefined}
-            data-active={session.isFlagged}
-            data-pending={pendingAction === 'flag' ? 'true' : undefined}
-            disabled={actionBusy}
-            title={session.isFlagged ? '取消置顶对话' : '置顶对话'}
-          >
-            {session.isFlagged
-              ? <PinOff size={14} strokeWidth={1.75} aria-hidden="true" />
-              : <Pin size={14} strokeWidth={1.75} aria-hidden="true" />}
-          </UiButton>
-          <UiButton
-            type="button"
-            variant="quiet"
-            size="icon-sm"
-            className="maka-list-row-action"
-            tabIndex={actionTabIndex}
-            onClick={startRename}
-            aria-label="重命名对话"
-            aria-busy={pendingAction === 'rename' ? 'true' : undefined}
-            data-pending={pendingAction === 'rename' ? 'true' : undefined}
-            disabled={actionBusy}
-            title="重命名（双击行名也可）"
-          >
-            <Pencil size={14} strokeWidth={1.75} aria-hidden="true" />
-          </UiButton>
-          <UiButton
-            type="button"
-            variant="quiet"
-            size="icon-sm"
-            className="maka-list-row-action"
-            tabIndex={actionTabIndex}
-            onClick={(event) => {
-              stopPropagation(event);
-              runRowAction('archive', () => (
-                session.isArchived
-                  ? actions.onUnarchive(session.id)
-                  : actions.onArchive(session.id)
-              ));
-            }}
-            aria-label={session.isArchived ? '取消归档对话' : '归档对话'}
-            aria-busy={pendingAction === 'archive' ? 'true' : undefined}
-            data-pending={pendingAction === 'archive' ? 'true' : undefined}
-            disabled={actionBusy}
-            title={session.isArchived ? '取消归档' : '归档'}
-          >
-            {session.isArchived
-              ? <ArchiveRestore size={14} strokeWidth={1.75} aria-hidden="true" />
-              : <Archive size={14} strokeWidth={1.75} aria-hidden="true" />}
-          </UiButton>
-          <UiButton
-            type="button"
-            variant="quiet"
-            size="icon-sm"
-            className="maka-list-row-action maka-list-row-action-danger"
-            tabIndex={actionTabIndex}
-            onClick={handleDelete}
-            aria-label="删除对话"
-            aria-busy={pendingAction === 'delete' ? 'true' : undefined}
-            data-pending={pendingAction === 'delete' ? 'true' : undefined}
-            disabled={actionBusy}
-            title="删除"
-          >
-            <Trash2 size={14} strokeWidth={1.75} aria-hidden="true" />
-          </UiButton>
-        </div>
-      )}
-    </div>
-  );
-}
 
 interface PermissionModeMeta {
   label: string;
@@ -6014,1877 +4310,59 @@ export const Composer = forwardRef<
   );
 });
 
-// Mirror of runtime's LOAD_TOOLS_NAME. @maka/ui must not depend on @maka/runtime,
-// so the always-on group-activation connector's name is duplicated here as the
-// single hook for its friendly, locale-aware presentation. The pre-unification
-// name `load_tool` (PR #30) is also matched — it shipped and returns the same
-// `{ loaded: [...] }` shape, so replayed old sessions still render friendly.
-// `connect_tool_source` (PR #34) is intentionally NOT here: it never shipped and
-// its `{ tools: [...] }` result shape this card does not render.
-const CONNECTOR_TOOL_NAMES: ReadonlySet<string> = new Set(['load_tools', 'load_tool']);
-function isConnectorTool(name: string): boolean {
-  return CONNECTOR_TOOL_NAMES.has(name);
-}
-
-/** Locale-aware display name for the group-activation connector. */
-export function loadToolDisplayName(locale: UiLocale): string {
-  return locale === 'en' ? 'Load tools' : '加载工具组';
-}
-
-interface LoadToolResultDescription {
-  title: string;
-  countLabel: string;
-  toolsText: string;
-  footer: string;
-}
-
-/**
- * Turn a `load_tools` call + its thin `{ loaded: [...] }` result into friendly,
- * locale-aware card copy. Reads the group id from `group` (current) or the
- * historical `namespace` arg (`load_tool`, PR #30) so replayed old sessions
- * still render. Returns `null` when the result is not the expected shape (e.g. a
- * load failure, a text/error result) so the caller falls back to the generic
- * preview.
- */
-export function describeLoadToolResult(
-  args: unknown,
-  value: unknown,
-  locale: UiLocale,
-): LoadToolResultDescription | null {
-  const loaded = (value as { loaded?: unknown } | null | undefined)?.loaded;
-  if (!Array.isArray(loaded) || !loaded.every((name) => typeof name === 'string')) {
-    return null;
-  }
-  const tools = loaded as string[];
-  const argRecord = args as { group?: unknown; namespace?: unknown } | null | undefined;
-  const rawGroup = argRecord?.group ?? argRecord?.namespace;
-  const namespace =
-    typeof rawGroup === 'string' && rawGroup.length > 0 ? rawGroup : undefined;
-  const n = tools.length;
-  if (locale === 'en') {
-    return {
-      title: namespace ? `Loaded ${namespace} tool group` : 'Tools loaded',
-      countLabel: n === 1 ? '1 tool now available:' : `${n} tools now available:`,
-      toolsText: tools.join(', '),
-      footer: 'Ready to use on the next step',
-    };
-  }
-  return {
-    title: namespace ? `已加载 ${namespace} 工具组` : '已加载工具组',
-    countLabel: `新增 ${n} 个可用工具：`,
-    toolsText: tools.join('、'),
-    footer: '下一步即可调用',
-  };
-}
-
-/** Friendly tool name: an explicit displayName wins; the connector gets a localized name. */
-function resolveToolDisplayName(item: ToolActivityItem): string {
-  if (item.displayName) return item.displayName;
-  if (isConnectorTool(item.toolName)) return loadToolDisplayName(detectUiLocale());
-  return item.toolName;
-}
-
-/** Friendly card for a `load_tools` result; falls back to JSON on unexpected shapes. */
-function LoadToolResultPreview(props: { args: unknown; value: unknown }) {
-  const desc = describeLoadToolResult(props.args, props.value, detectUiLocale());
-  if (!desc) {
-    return <OverlayPreview content={{ kind: 'json', value: props.value }} />;
-  }
-  return (
-    <div className={previewVariants({ part: 'load-tool' })} data-kind="load_tool">
-      <p className={previewVariants({ part: 'load-tool-title' })}>{desc.title}</p>
-      <p className={previewVariants({ part: 'load-tool-count' })}>{desc.countLabel}</p>
-      <p className={previewVariants({ part: 'load-tool-tools' })}>{desc.toolsText}</p>
-      <p className={previewVariants({ part: 'load-tool-footer' })}>{desc.footer}</p>
-    </div>
-  );
-}
-
-const STATUS_LABEL: Record<ToolActivityItem['status'], string> = {
-  pending: '排队中',
-  waiting_permission: '等待权限',
-  running: '运行中',
-  completed: '已完成',
-  errored: '失败',
-  interrupted: '已中断',
-};
-
-function isOpenByDefault(status: ToolActivityItem['status']): boolean {
-  // Show details inline while the call is in flight or blocking the user; also
-  // for errored calls so the failure is visible without an extra click. Settled
-  // success / interruption collapse so completed history doesn't drown the chat.
-  return (
-    status === 'pending' ||
-    status === 'waiting_permission' ||
-    status === 'running' ||
-    status === 'errored'
-  );
-}
-
-function extractErrorText(result: ToolActivityItem['result']): string {
-  if (!result) return '';
-  switch (result.kind) {
-    case 'text':
-      return result.text;
-    case 'json':
-      try {
-        return JSON.stringify(result.value, null, 2);
-      } catch {
-        return String(result.value);
-      }
-    case 'terminal':
-      return result.stderr || result.stdout || `exit ${result.exitCode}`;
-    case 'file_diff':
-      return result.diff;
-    case 'rive_workflow':
-      return result.error
-        ? [result.summary, result.error.reason, result.error.message].filter(Boolean).join('\n')
-        : result.summary;
-    default:
-      return result.kind;
-  }
-}
-
-function formatUserVisibleToolText(text: string): string {
-  return text.replace(/\bUser denied permission\b/g, '用户已拒绝权限请求');
-}
-
-function isPermissionDeniedToolResult(result: ToolActivityItem['result']): boolean {
-  return result?.kind === 'text' && formatUserVisibleToolText(result.text).trim() === '用户已拒绝权限请求';
-}
-
-export function formatRedactedJson(value: unknown): string {
-  try {
-    return redactSecrets(JSON.stringify(value, null, 2));
-  } catch {
-    return redactSecrets(String(value));
-  }
-}
-
-export function formatToolIntent(intent: string): string {
-  const safe = redactSecrets(intent.replace(/\s+/g, ' ').trim());
-  return safe.length > 240 ? `${safe.slice(0, 240)}…` : safe;
-}
-
-function formatDuration(ms: number | undefined): string | null {
-  if (ms === undefined || ms < 0) return null;
-  if (ms < 1000) return `${ms} ms`;
-  if (ms < 60_000) return `${(ms / 1000).toFixed(ms < 10_000 ? 1 : 0)}s`;
-  const minutes = Math.floor(ms / 60_000);
-  const seconds = Math.round((ms % 60_000) / 1000);
-  return `${minutes}m ${seconds}s`;
-}
-
-export function ToolActivity(props: { items: ToolActivityItem[] }) {
-  return (
-    <section className={toolVariants({ part: 'container' })} aria-label="工具调用记录">
-      <header className={toolVariants({ part: 'container-header' })}>
-        <strong>工具调用</strong>
-        <span className={toolVariants({ part: 'count' })} aria-label={`${props.items.length} 次调用`}>{props.items.length}</span>
-      </header>
-      {props.items.map((item) => {
-        const duration = formatDuration(item.durationMs);
-        const errored = item.status === 'errored';
-        const permissionDenied = isPermissionDeniedToolResult(item.result);
-        return (
-          <details
-            key={item.toolUseId}
-            data-slot="tool"
-            className={toolVariants({ part: 'item' })}
-            data-status={item.status}
-            open={isOpenByDefault(item.status)}
-          >
-            <summary className={toolVariants({ part: 'header' })}>
-              <span className={toolVariants({ part: 'dot' })} data-status={item.status} aria-hidden="true" />
-              <span className={toolVariants({ part: 'name' })}>{resolveToolDisplayName(item)}</span>
-              <span className={toolVariants({ part: 'meta' })}>
-                {duration && <span className={toolVariants({ part: 'duration' })}>{duration}</span>}
-                <span className={toolVariants({ part: 'status-label' })}>{STATUS_LABEL[item.status]}</span>
-              </span>
-            </summary>
-            <div className={toolVariants({ part: 'body' })}>
-              {errored && <ToolErrorBanner result={item.result} />}
-              {item.intent && !permissionDenied && <p className={toolVariants({ part: 'intent' })}>{formatToolIntent(item.intent)}</p>}
-              {item.args !== undefined && !permissionDenied && (
-                <pre className={`maka-code ${toolVariants({ part: 'args' })}`}>{formatRedactedJson(item.args)}</pre>
-              )}
-              {item.outputChunks && item.outputChunks.length > 0 && (
-                <ToolOutputStream
-                  chunks={item.outputChunks}
-                  live={item.status === 'running' || item.status === 'pending'}
-                  interrupted={item.status === 'interrupted'}
-                  truncated={item.outputTruncated === true}
-                />
-              )}
-              {item.result && !permissionDenied && (
-                isConnectorTool(item.toolName) && item.result.kind === 'json' ? (
-                  <LoadToolResultPreview args={item.args} value={item.result.value} />
-                ) : (
-                  <OverlayPreview content={item.result} />
-                )
-              )}
-            </div>
-          </details>
-        );
-      })}
-    </section>
-  );
-}
-
-/**
- * PR-UI-12 — live stdout/stderr stream from PR-REAL-4 `tool_output_delta`.
- *
- * Renders chunks in their original seq order (already sorted in main.tsx
- * before this component sees them) so interleaved stdout+stderr reads
- * the way a human would expect from a real terminal. Each chunk keeps
- * its stream tag so stderr can render in a destructive tone — a
- * single mono `<pre>` would lose that visual signal.
- *
- * `redacted: true` chunks render as a small inline hint "[已脱敏]"
- * instead of pretending the chunk arrived clean. Empty redacted
- * chunks (runtime suppressed everything) collapse to just the hint.
- *
- * `truncated: true` (PR-UI-12 fixup #2, @kenji A3 msg 365ff8b9) flips
- * a "已截断" pill in the header counts row. This means
- * `applyToolOutputChunk` dropped chunks (per-tool count or
- * total-char cap) or tail-truncated a single oversize chunk. Users
- * see explicitly that the displayed stream is bounded — they should
- * use Finder / external viewer if they need the full output.
- *
- * Auto-scroll: while `live` is true, we anchor to the bottom on every
- * chunk update so users see the latest output. Once the tool reaches
- * terminal (`tool_result`), auto-scroll stops so users can scroll up
- * to read history without being yanked back.
- */
-function ToolOutputStream(props: {
-  chunks: ToolOutputChunk[];
-  live: boolean;
-  interrupted: boolean;
-  truncated: boolean;
-}) {
-  const preRef = useRef<HTMLPreElement>(null);
-  useEffect(() => {
-    if (!props.live) return;
-    const el = preRef.current;
-    if (!el) return;
-    el.scrollTop = el.scrollHeight;
-  }, [props.chunks, props.live]);
-
-  const stdoutCount = props.chunks.filter((c) => c.stream === 'stdout').length;
-  const stderrCount = props.chunks.filter((c) => c.stream === 'stderr').length;
-  const redactedCount = props.chunks.filter((c) => c.redacted).length;
-
-  return (
-    <div className={streamVariants({ part: 'container' })} data-live={props.live ? 'true' : undefined}>
-      <header className={streamVariants({ part: 'header' })}>
-        <span className={streamVariants({ part: 'label' })}>
-          {props.live ? (
-            <>
-              <LiveIndicator />
-              <span>实时输出</span>
-            </>
-          ) : props.interrupted ? (
-            <span>已中断 · 已收到的输出</span>
-          ) : (
-            <span>工具输出</span>
-          )}
-        </span>
-        <span className={streamVariants({ part: 'counts' })}>
-          {stdoutCount > 0 && <span className={streamVariants({ part: 'count' })}>stdout {stdoutCount}</span>}
-          {stderrCount > 0 && <span className={streamVariants({ part: 'count' })} data-stream="stderr">stderr {stderrCount}</span>}
-          {redactedCount > 0 && <span className={streamVariants({ part: 'count' })} data-redacted="true">已脱敏 {redactedCount}</span>}
-          {props.truncated && (
-            <span
-              className={streamVariants({ part: 'count' })}
-              data-truncated="true"
-              title="部分输出已截断；如需完整输出请查看对应工具结果或生成的 artifact"
-            >
-              已截断
-            </span>
-          )}
-        </span>
-      </header>
-      <pre ref={preRef} className={streamVariants({ part: 'body' })}>
-        {props.chunks.map((chunk) => (
-          <span
-            key={chunk.seq}
-            className={streamVariants({ part: 'chunk' })}
-            data-stream={chunk.stream}
-            data-redacted={chunk.redacted ? 'true' : undefined}
-          >
-            {chunk.text}
-            {chunk.redacted && (
-              <span className={streamVariants({ part: 'redacted-tag' })} aria-label="已脱敏">
-                {' '}[已脱敏]
-              </span>
-            )}
-          </span>
-        ))}
-      </pre>
-    </div>
-  );
-}
-
-// Preserve the retired `.maka-tool-error*` leaf utilities onto Alert (#332 PR3c) —
-// Alert owns the shell; these are the few declarations it doesn't set, kept arbitrary
-// so they map 1:1 to the old CSS (`[align-self:start]`, not Tailwind's `flex-start`).
-function ToolErrorBanner(props: { result: ToolActivityItem['result'] }) {
-  // Tool stderr / raw provider errors occasionally slip credential paths,
-  // bearer tokens, or API keys through main-side redaction. Apply a
-  // defensive UI-level mask before display *and* before clipboard copy so
-  // the user can't accidentally paste a credential into a bug report.
-  const errorText = formatUserVisibleToolText(redactSecrets(extractErrorText(props.result)));
-  const copyFeedback = useClipboardCopyFeedback();
-  const copyPhase = copyFeedback.phaseFor('tool-error');
-  const copyPending = copyPhase === 'pending';
-  const copyLabel = copyPhase === 'pending'
-    ? '复制中…'
-    : copyPhase === 'copied'
-      ? '已复制'
-      : copyPhase === 'failed'
-        ? '复制失败'
-        : '复制';
-
-  async function copy() {
-    if (!errorText) return;
-    await copyFeedback.copy('tool-error', errorText);
-  }
-
-  return (
-    <Alert variant="error" className="mb-[10px]">
-      <AlertOctagon size={16} strokeWidth={2} aria-hidden="true" />
-      <AlertTitle>工具调用失败</AlertTitle>
-      {errorText && (
-        <AlertDescription className="[font-family:var(--font-mono)] text-[12px] leading-[1.5] whitespace-pre-wrap [word-break:break-word]">
-          {errorText.length > 240 ? `${errorText.slice(0, 240)}…` : errorText}
-        </AlertDescription>
-      )}
-      {errorText && (
-        <AlertAction>
-          <UiButton
-            type="button"
-            variant="ghost"
-            size="sm"
-            className="maka-button [align-self:start] data-[pending=true]:cursor-progress data-[copy-feedback=copied]:text-[color:var(--accent)] data-[copy-feedback=copied]:border-[oklch(from_var(--accent)_l_c_h_/_0.35)] data-[copy-feedback=failed]:text-[color:var(--destructive)] data-[copy-feedback=failed]:border-[oklch(from_var(--destructive)_l_c_h_/_0.35)]"
-            data-pending={copyPending ? 'true' : undefined}
-            data-copy-feedback={copyPhase ?? undefined}
-            aria-label={`${copyLabel}错误信息`}
-            aria-busy={copyPending ? 'true' : undefined}
-            disabled={copyPending}
-            onClick={() => void copy()}
-          >
-            {copyPhase === 'copied' ? <Check size={14} aria-hidden="true" /> : <Copy size={14} aria-hidden="true" />}
-            <span>{copyLabel}</span>
-          </UiButton>
-        </AlertAction>
-      )}
-    </Alert>
-  );
-}
-
-export function OverlayHost(props: { content?: ToolResultContent; onClose(): void }) {
-  if (!props.content) return null;
-  return (
-    <div className="maka-modal-backdrop overlay">
-      <UiButton
-        className={cn('maka-button', previewVariants({ part: 'close' }))}
-        type="button"
-        variant="ghost"
-        onClick={props.onClose}
-        aria-label="关闭预览"
-      >
-        <X size={14} strokeWidth={1.75} aria-hidden="true" />
-        <span>关闭</span>
-      </UiButton>
-      <OverlayPreview content={props.content} />
-    </div>
-  );
-}
-
-// Per-reason presentation hints. Drives icon + headline + risk tone in the
-// dialog so the user can scan the modal in 1-2 seconds before reading the
-// args block.
-type ReasonKind = PermissionRequestEvent['reason'];
-
-interface ReasonPreset {
-  label: string;
-  Icon: typeof AlertTriangle;
-  tone: 'info' | 'caution' | 'destructive';
-}
-
-const REASON_PRESETS: Record<ReasonKind, ReasonPreset> = {
-  shell_dangerous: { label: '高风险 shell 命令 · 请仔细确认', Icon: Terminal, tone: 'caution' },
-  file_write: { label: '写入或创建文件', Icon: FileEdit, tone: 'info' },
-  fs_destructive: { label: '不可恢复的文件系统操作', Icon: AlertOctagon, tone: 'destructive' },
-  git_destructive: { label: '不可恢复的 Git 操作', Icon: GitMerge, tone: 'destructive' },
-  network: { label: '对外网络请求', Icon: Wifi, tone: 'info' },
-  privileged: { label: '特权操作 (sudo / su)', Icon: ShieldAlert, tone: 'destructive' },
-  browser: { label: '读取和操作你登录的浏览器会话 · 请确认', Icon: Globe, tone: 'caution' },
-  custom: { label: '自定义请求', Icon: HelpCircle, tone: 'info' },
-};
-
-export function PermissionDialog(props: {
-  request: PermissionRequestEvent;
-  // Accept Promise-returning impls so the dialog can await the IPC
-  // and reset its own pending state when it resolves OR rejects.
-  // The renderer's `respondToPermission` is async but was typed as
-  // void by the legacy signature, which made `submit()` strand
-  // `responsePending=true` if the IPC failed silently.
-  onRespond(response: PermissionResponse): void | Promise<void>;
-}) {
-  const [rememberForTurn, setRememberForTurn] = useState(false);
-  const [responsePending, setResponsePending] = useState(false);
-  const [now, setNow] = useState(() => Date.now());
-  const dialogRef = useRef<HTMLDivElement>(null);
-  const responsePendingRef = useRef(false);
-  const permissionMountedRef = useRef(true);
-  const activePermissionRequestIdRef = useRef(props.request.requestId);
-  // No onEscape — a permission request requires an explicit allow/deny decision.
-  useModalA11y(dialogRef);
-
-  useEffect(() => {
-    permissionMountedRef.current = true;
-    return () => {
-      permissionMountedRef.current = false;
-    };
-  }, []);
-
-  useEffect(() => {
-    activePermissionRequestIdRef.current = props.request.requestId;
-    setRememberForTurn(false);
-    setResponsePending(false);
-    responsePendingRef.current = false;
-    setNow(Date.now());
-  }, [props.request.requestId]);
-
-  useEffect(() => {
-    const tick = () => setNow(Date.now());
-    const interval = window.setInterval(tick, 30_000);
-    return () => window.clearInterval(interval);
-  }, [props.request.requestId]);
-
-  async function submit(decision: PermissionResponse['decision']) {
-    if (responsePendingRef.current) return;
-    const requestId = props.request.requestId;
-    responsePendingRef.current = true;
-    setResponsePending(true);
-    try {
-      // PR-PERMISSION-UI-CLEANUP-0: await so the pending state
-      // resets when the IPC settles. Previously a Promise-returning
-      // onRespond would let the try/catch miss async rejections,
-      // and on success the parent normally unmounts us — but if the
-      // parent's own try/catch swallows the IPC error (PR-STOP-
-      // ERROR-SURFACE-0 does exactly this), we'd stay mounted with
-      // `responsePending=true` and the buttons would lock up.
-      await props.onRespond({
-        requestId,
-        decision,
-        rememberForTurn: decision === 'allow' ? rememberForTurn : false,
-      });
-    } finally {
-      if (activePermissionRequestIdRef.current === requestId) {
-        responsePendingRef.current = false;
-        if (permissionMountedRef.current) setResponsePending(false);
-      }
-    }
-  }
-
-  const preset = REASON_PRESETS[props.request.reason] ?? REASON_PRESETS.custom;
-  const summary = renderPermissionSummary(props.request);
-  const isDestructive = preset.tone === 'destructive';
-  const health = derivePermissionRequestHealth({ requestedAt: props.request.ts, now });
-  const waitLabel = formatPermissionRequestWait(health.ageMs);
-
-  return (
-    <DialogRoot open disablePointerDismissal>
-      <DialogContent
-        ref={dialogRef}
-        className="maka-modal permissionDialog w-[min(92vw,720px)] p-0"
-        role="alertdialog"
-        aria-labelledby="permissionTitle"
-        data-tone={preset.tone}
-        showClose={false}
-      >
-        <div className="maka-modal-header maka-permission-header">
-          <span className="maka-permission-icon" aria-hidden="true">
-            <preset.Icon size={20} strokeWidth={1.75} />
-          </span>
-          <div>
-            <h2 className="maka-modal-title" id="permissionTitle">需要确认权限</h2>
-            <p className="maka-modal-subtitle">
-              <Badge variant={isDestructive ? 'destructive' : 'secondary'} className="maka-permission-tool font-mono">{props.request.toolName}</Badge>
-              <span aria-hidden="true"> · </span>
-              <span className="maka-reason-text" data-reason={props.request.reason}>{preset.label}</span>
-              <span aria-hidden="true"> · </span>
-              <span className="maka-permission-age" data-status={health.status}>
-                已等待 {waitLabel}
-              </span>
-            </p>
-          </div>
-        </div>
-        <div className="maka-modal-body maka-permission-body">
-          {summary && <div className="maka-permission-summary">{summary}</div>}
-          {props.request.hint && (
-            <div className="maka-permission-hint" role="note">{props.request.hint}</div>
-          )}
-          <details className="maka-permission-raw">
-            <summary>查看完整参数</summary>
-            <pre className="maka-code">{formatRedactedJson(props.request.args)}</pre>
-          </details>
-          <label className="permissionRemember">
-            <Checkbox
-              checked={rememberForTurn}
-              disabled={responsePending}
-              onCheckedChange={(checked) => setRememberForTurn(checked === true)}
-            />
-            本轮对话内记住选择（同类型工具不再询问，关闭/切换对话后失效）
-          </label>
-          {props.request.reason === 'browser' && rememberForTurn && (
-            <p className="maka-permission-hint" role="note">
-              勾选后，本轮接下来的浏览、读取页面、导航、点击、输入都不再逐次询问。你会全程看到它操作的页面，随时可以停止；本轮结束后授权失效。
-            </p>
-          )}
-          {isDestructive && (
-            <Alert variant="error" className="maka-permission-danger-note">
-              <AlertDescription>
-                这类操作不可恢复，确认前请再读一遍上面的参数。
-              </AlertDescription>
-            </Alert>
-          )}
-          {health.status !== 'fresh' && (
-            <Alert
-              variant="warning"
-              className="maka-permission-stale-note"
-              data-status={health.status}
-            >
-              <AlertDescription>
-                这个请求已经等待较久。允许前请重新确认工具名和参数；如果上下文已经变了，直接拒绝后重新发送。
-              </AlertDescription>
-            </Alert>
-          )}
-        </div>
-        <div className="maka-modal-footer permissionActions">
-          <UiButton className="maka-button" variant="ghost" type="button" disabled={responsePending} onClick={() => submit('deny')}>拒绝</UiButton>
-          <UiButton
-            className="maka-button"
-            variant={isDestructive ? 'destructive' : 'default'}
-            type="button"
-            disabled={responsePending}
-            onClick={() => submit('allow')}
-          >
-            {responsePending ? '正在提交…' : isDestructive ? '我已确认，允许' : '允许'}
-          </UiButton>
-        </div>
-      </DialogContent>
-    </DialogRoot>
-  );
-}
-
-/**
- * One-line summary for a browser_* action. Names the concrete action (open /
- * read / click / type) so the prompt reads as a real browser step, not an opaque
- * tool call — reinforcing that a browser grant spans reads AND acts. The typed
- * text and full args stay in the raw `<details>` block below.
- */
-function renderBrowserSummary(toolName: string, args: Record<string, unknown>): ReactNode {
-  const ref = typeof args.ref === 'string' ? args.ref : '';
-  const url = typeof args.url === 'string' ? args.url : '';
-  const selector = typeof args.selector === 'string' ? args.selector : '';
-  const line =
-    toolName === 'browser_navigate'
-      ? `即将在浏览器中打开 ${url || '一个网址'}`
-      : toolName === 'browser_click'
-        ? `即将在当前页面点击元素 ${ref}`.trim()
-        : toolName === 'browser_type'
-          ? `即将在当前页面输入文本${ref ? ` 到元素 ${ref}` : ''}`
-          : toolName === 'browser_snapshot'
-            ? '即将读取当前页面的可交互元素列表'
-            : toolName === 'browser_extract'
-              ? `即将读取当前页面内容${selector ? `（${selector}）` : ''}`
-              : toolName === 'browser_wait'
-                ? '即将等待当前页面满足某个条件'
-                : '即将操作当前浏览器页面';
-  return <p className="maka-permission-line">{line}</p>;
-}
-
-/**
- * Per-tool human-readable summary of what the request will do, used at the
- * top of the permission dialog body. Falls back to undefined if we can't
- * recognize the tool — the raw args `<details>` block is always available.
- */
-function renderPermissionSummary(request: PermissionRequestEvent): ReactNode | undefined {
-  const args = (request.args ?? {}) as Record<string, unknown>;
-  switch (request.toolName) {
-    case 'browser_navigate':
-    case 'browser_snapshot':
-    case 'browser_click':
-    case 'browser_type':
-    case 'browser_wait':
-    case 'browser_extract':
-      return renderBrowserSummary(request.toolName, args);
-    case 'Bash': {
-      const command = typeof args.command === 'string' ? args.command : undefined;
-      if (!command) return undefined;
-      const timeout = typeof args.timeout_ms === 'number' ? args.timeout_ms : undefined;
-      return (
-        <>
-          <p className="maka-permission-line">即将运行 shell 命令：</p>
-          <pre className="maka-code maka-permission-command">{redactSecrets(command)}</pre>
-          {timeout !== undefined && (
-            <p className="maka-permission-meta">超时 <strong>{timeout} ms</strong></p>
-          )}
-        </>
-      );
-    }
-    case 'Write': {
-      const path = typeof args.path === 'string' ? args.path : undefined;
-      const content = typeof args.content === 'string' ? args.content : '';
-      if (!path) return undefined;
-      const bytes = new TextEncoder().encode(content).length;
-      const lines = content.split('\n').length;
-      const preview = permissionTextPreview(content, 600);
-      return (
-        <>
-          <p className="maka-permission-line">即将写入文件：</p>
-          <p className="maka-permission-path"><code>{redactSecrets(path)}</code></p>
-          <p className="maka-permission-meta">
-            <strong>{bytes}</strong> 字节 · <strong>{lines}</strong> 行
-          </p>
-          <pre className="maka-code maka-permission-preview">{preview}</pre>
-        </>
-      );
-    }
-    case 'Edit': {
-      const path = typeof args.path === 'string' ? args.path : undefined;
-      const oldString = typeof args.old_string === 'string' ? args.old_string : '';
-      const newString = typeof args.new_string === 'string' ? args.new_string : '';
-      if (!path) return undefined;
-      return (
-        <>
-          <p className="maka-permission-line">即将修改文件：</p>
-          <p className="maka-permission-path"><code>{redactSecrets(path)}</code></p>
-          <div className="maka-permission-diff">
-            <div>
-              <span className="maka-permission-diff-tag" data-side="old">删除</span>
-              <pre className="maka-code">{permissionTextPreview(oldString, 400)}</pre>
-            </div>
-            <div>
-              <span className="maka-permission-diff-tag" data-side="new">写入</span>
-              <pre className="maka-code">{permissionTextPreview(newString, 400)}</pre>
-            </div>
-          </div>
-        </>
-      );
-    }
-    case 'OfficeDocumentEdit': {
-      const path = typeof args.path === 'string' ? args.path : undefined;
-      const operation = typeof args.operation === 'string' ? args.operation : undefined;
-      if (!path || !operation) return undefined;
-      const target = typeof args.target === 'string' ? args.target : undefined;
-      const elementType = typeof args.elementType === 'string' ? args.elementType : undefined;
-      const index = typeof args.index === 'number' ? args.index : undefined;
-      const propsArg = args.props && typeof args.props === 'object' && !Array.isArray(args.props)
-        ? args.props as Record<string, unknown>
-        : {};
-      const propEntries = Object.entries(propsArg).slice(0, 6);
-      const hiddenProps = Math.max(0, Object.keys(propsArg).length - propEntries.length);
-      return (
-        <>
-          <p className="maka-permission-line">即将编辑 Office 文档：</p>
-          <p className="maka-permission-path"><code>{redactSecrets(path)}</code></p>
-          <p className="maka-permission-meta">
-            操作 <strong>{redactSecrets(operation)}</strong>
-            {target && <> · 目标 <code>{redactSecrets(target)}</code></>}
-            {elementType && <> · 元素 <code>{redactSecrets(elementType)}</code></>}
-            {index !== undefined && <> · 位置 <strong>{index}</strong></>}
-          </p>
-          {propEntries.length > 0 && (
-            <pre className="maka-code maka-permission-preview">
-              {propEntries.map(([key, value]) => `${redactSecrets(key)}=${permissionValuePreview(value)}`).join('\n')}
-              {hiddenProps > 0 && `\n… 另有 ${hiddenProps} 个属性`}
-            </pre>
-          )}
-        </>
-      );
-    }
-    default:
-      return undefined;
-  }
-}
-
-function permissionTextPreview(value: string, maxChars: number): string {
-  const safe = redactSecrets(value);
-  return safe.length > maxChars ? `${safe.slice(0, maxChars)}…` : safe;
-}
-
-function permissionValuePreview(value: unknown): string {
-  if (typeof value === 'string') {
-    const safe = redactSecrets(value);
-    return safe.length > 160 ? `${safe.slice(0, 160)}…` : safe;
-  }
-  if (typeof value === 'number' && Number.isFinite(value)) return String(value);
-  if (typeof value === 'boolean') return value ? 'true' : 'false';
-  return '不支持的属性值';
-}
-
-/**
- * Renders a ToolResultContent payload with kind-specific presentation:
- * - `file_diff`: line-level red/green diff coloring
- * - `terminal`: stdout + stderr split with exit-code badge + stderr in
- *   destructive tone
- * - `office_document`: Office adapter stdout/stderr/diagnostic cards
- * - `explore_agent`: bounded read-only subagent findings
- * - `subagent`: foreground child-agent run summary
- * - `json`: pretty-printed in a code block
- * - `text` / others: plain `<pre>` fallback
- *
- * All variants are height-bounded by the `@maka/ui` previewVariants `overlay`
- * part (the retired `.maka-overlay-preview` base) to keep kilobyte outputs from
- * pushing the composer off-screen.
- */
-/**
- * Cap displayed line count to keep a giant tool output (10k-line stderr from
- * a failing test run) from creating 10k React elements and from drowning the
- * chat surface visually. We slice, then append a single explainer line that
- * lets the user know the rest exists.
- */
-const TOOL_LINE_CAP = 500;
-
-function capLines(text: string): { body: string; capped: number } {
-  const lines = text.split('\n');
-  if (lines.length <= TOOL_LINE_CAP) return { body: text, capped: 0 };
-  return {
-    body: lines.slice(0, TOOL_LINE_CAP).join('\n'),
-    capped: lines.length - TOOL_LINE_CAP,
-  };
-}
-
-function OverlayPreview(props: { content: ToolResultContent }) {
-  const { content } = props;
-
-  if (content.kind === 'file_diff') {
-    return <FileDiffPreview diff={content.diff} paths={content.paths} />;
-  }
-
-  if (content.kind === 'web_search') {
-    return (
-      <WebSearchPreview query={content.query} provider={content.provider} rows={content.rows} />
-    );
-  }
-
-  if (content.kind === 'web_search_error') {
-    return (
-      <WebSearchErrorPreview
-        query={content.query}
-        provider={content.provider}
-        reason={content.reason}
-        message={content.message}
-        credentialSource={content.credentialSource}
-      />
-    );
-  }
-
-  if (content.kind === 'terminal') {
-    return (
-      <TerminalPreview
-        cwd={content.cwd}
-        cmd={content.cmd}
-        exitCode={content.exitCode}
-        stdout={content.stdout}
-        stderr={content.stderr}
-      />
-    );
-  }
-
-  if (content.kind === 'office_document') {
-    return <OfficeDocumentPreview result={content} />;
-  }
-
-  if (content.kind === 'explore_agent') {
-    return <ExploreAgentPreview result={content} />;
-  }
-
-  if (content.kind === 'subagent') {
-    return <SubagentPreview result={content} />;
-  }
-
-  if (content.kind === 'rive_workflow') {
-    return <RiveWorkflowPreview result={content} />;
-  }
-
-  if (content.kind === 'json') {
-    let body: string;
-    try {
-      body = JSON.stringify(content.value, null, 2);
-    } catch {
-      body = String(content.value);
-    }
-    // JSON shouldn't contain secrets persisted by Maka (settings + telemetry
-    // are sanitized at write-time), but apply the renderer redactor as a
-    // second-layer defense in case a tool returned raw provider response.
-    return <pre className={previewVariants({ part: 'overlay' })} data-kind="json">{formatUserVisibleToolText(redactSecrets(body))}</pre>;
-  }
-
-  if (content.kind === 'text') {
-    const { body, capped } = capLines(formatUserVisibleToolText(redactSecrets(content.text)));
-    return (
-      <pre className={previewVariants({ part: 'overlay' })} data-kind="text">
-        {body}
-        {capped > 0 && `\n\n… 已隐藏 ${capped} 行`}
-      </pre>
-    );
-  }
-
-  // file_write / image / summary / unknown — show a compact descriptor so the
-  // user knows what kind landed without dumping binary or storage refs.
-  return (
-    <pre className={previewVariants({ part: 'overlay' })} data-kind={content.kind}>
-      [{content.kind}]
-    </pre>
-  );
-}
-
-function RiveWorkflowPreview(props: {
-  result: Extract<ToolResultContent, { kind: 'rive_workflow' }>;
-}) {
-  const { result } = props;
-  const rows = [
-    ['动作', result.action],
-    ['状态', result.state ?? result.projection?.state],
-    ['workflow_run', result.ids.workflowRunId ?? result.projection?.workflowRunId],
-    ['scheduler_run', result.ids.schedulerRunId ?? result.projection?.schedulerRunId],
-    ['root_work', result.ids.rootWorkNodeId ?? result.projection?.rootWorkNodeId],
-    ['scheduler_state', result.projection?.schedulerState],
-    ['root_state', result.projection?.rootState],
-  ].filter((row): row is [string, string] => typeof row[1] === 'string' && row[1].length > 0);
-  const nodes = (result.nodes ?? []).slice(0, 12);
-  const failureLines = result.error
-    ? [
-        '',
-        '错误',
-        `reason: ${result.error.reason}`,
-        `message: ${result.error.message}`,
-        result.error.code ? `code: ${result.error.code}` : '',
-        result.error.suggestedAction ? `suggested_action: ${result.error.suggestedAction}` : '',
-      ].filter(Boolean)
-    : [];
-  const diagnosticLines = [
-    result.stdoutTail ? `stdout_tail:\n${redactSecrets(result.stdoutTail)}` : '',
-    result.stderrTail ? `stderr_tail:\n${redactSecrets(result.stderrTail)}` : '',
-  ].filter(Boolean);
-  const body = [
-    result.ok ? 'Rive workflow completed' : 'Rive workflow failed',
-    result.summary,
-    '',
-    ...rows.map(([label, value]) => `${label}: ${value}`),
-    ...(nodes.length > 0 ? ['', '节点摘要', ...nodes.map(formatRiveWorkflowNode)] : []),
-    ...failureLines,
-    ...(diagnosticLines.length > 0 ? ['', '诊断片段', ...diagnosticLines] : []),
-  ].join('\n');
-  const cappedPreview = capLines(body);
-  return (
-    <pre className={previewVariants({ part: 'overlay' })} data-kind="rive_workflow">
-      {cappedPreview.body}
-      {cappedPreview.capped > 0 && `\n\n… 已隐藏 ${cappedPreview.capped} 行`}
-    </pre>
-  );
-}
-
-function formatRiveWorkflowNode(node: NonNullable<Extract<ToolResultContent, { kind: 'rive_workflow' }>['nodes']>[number]): string {
-  const label = node.title ?? node.templateId ?? node.id ?? 'node';
-  const attrs = [
-    node.state,
-    node.runner ? `runner=${node.runner}` : '',
-    node.worker ? `worker=${node.worker}` : '',
-  ].filter(Boolean).join(' · ');
-  return attrs ? `- ${label}: ${attrs}` : `- ${label}`;
-}
-
-type SubagentResult = Extract<ToolResultContent, { kind: 'subagent' }>;
-
-const SUBAGENT_STATUS_LABEL: Record<SubagentResult['status'], string> = {
-  completed: '已完成',
-  failed: '失败',
-  cancelled: '已取消',
-  running: '运行中',
-  waiting_permission: '等待权限',
-};
-
-function SubagentPreview(props: {
-  result: SubagentResult;
-}) {
-  const { result } = props;
-  const duration = formatDuration(result.durationMs);
-  const status = presentSubagentStatus(result.status);
-  const summary = typeof result.summary === 'string' ? result.summary.trim() : '';
-  const artifactCount = result.artifactIds.length;
-  const meta = [
-    status,
-    presentSubagentPermission(result.permissionMode),
-    duration ? `耗时 ${duration}` : '',
-  ].filter(Boolean).join(' · ');
-
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'agent' }))} data-kind="subagent" data-status={result.status}>
-      <header className={previewVariants({ part: 'agent-head' })}>
-        <strong>{redactSecrets(result.agentName || 'Subagent')}</strong>
-        <small>{meta}</small>
-      </header>
-      {summary.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="子代理结果摘要">
-          <strong>结果摘要</strong>
-          <p>{redactSecrets(summary)}</p>
-        </section>
-      )}
-      {result.failureClass && (
-        <div className={previewVariants({ part: 'agent-message' })} role="note">
-          {redactSecrets(result.failureClass)}
-        </div>
-      )}
-      {artifactCount > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="子代理产物">
-          <strong>产物</strong>
-          <p>{artifactCount} 个</p>
-        </section>
-      )}
-    </div>
-  );
-}
-
-function presentSubagentStatus(status: SubagentResult['status']): string {
-  return SUBAGENT_STATUS_LABEL[status] ?? status;
-}
-
-function presentSubagentPermission(permissionMode: SubagentResult['permissionMode']): string {
-  if (permissionMode === 'explore') return '只读';
-  return permissionMode;
-}
-
-function ExploreAgentPreview(props: {
-  result: Extract<ToolResultContent, { kind: 'explore_agent' }>;
-}) {
-  const { result } = props;
-  const copyFeedback = useClipboardCopyFeedback();
-  const candidateFiles = result.candidateFiles.slice(0, 8);
-  const matches = result.matches.slice(0, 8);
-  const processLines = Array.isArray(result.recentEvents) && result.recentEvents.length > 0
-    ? result.recentEvents.slice(0, 20).map((event) => formatExploreAgentEvent(event, result.startedAt))
-    : (result.progress ?? []).slice(0, 12);
-  const progress = processLines.slice(0, 6);
-  const evidence = (result.evidence ?? []).slice(0, 6);
-  const resultSummary = typeof result.summary === 'string' ? result.summary.trim() : '';
-  const reportText = typeof result.report === 'string' ? result.report.trim() : '';
-  const terminalStatus = presentExploreAgentTerminalStatus(result.terminalStatus, result.ok, result.partial === true, result.reason);
-  const status = result.ok
-    ? '已完成'
-    : result.reason === 'aborted' && result.partial === true
-      ? '已取消 · 保留部分结果'
-      : presentExploreAgentReason(result.reason) ?? '未完成';
-  const reportLines = reportText.split('\n').filter((line) => line.trim().length > 0).slice(0, 8);
-  const notes = result.notes.slice(0, 4);
-  const roots = result.roots.length > 0 ? result.roots.join(', ') : '.';
-  const queries = result.queries.length > 0 ? result.queries.join(', ') : '未指定';
-  const ignoredPaths = Array.isArray(result.ignoredPaths) && result.ignoredPaths.length > 0
-    ? result.ignoredPaths.join(', ')
-    : '';
-  const stoppingCondition = typeof result.stoppingCondition === 'string'
-    ? result.stoppingCondition.trim()
-    : '';
-  const limitReasons = Array.isArray(result.limitReasons)
-    ? result.limitReasons.map(presentExploreAgentLimitReason).filter(Boolean).join('、')
-    : '';
-  const filesDiscovered = typeof result.filesDiscovered === 'number' && Number.isFinite(result.filesDiscovered)
-    ? Math.max(0, Math.floor(result.filesDiscovered))
-    : result.filesInspected;
-  const skippedSummary = result.sensitiveFilesSkipped && result.sensitiveFilesSkipped > 0
-    ? `跳过 ${result.filesSkipped} 个（含敏感 ${result.sensitiveFilesSkipped} 个）`
-    : `跳过 ${result.filesSkipped} 个`;
-  const duration = formatDuration(result.durationMs);
-  const summaryText = resultSummary.length > 0
-    ? [
-      `状态：${status}`,
-      `终态：${terminalStatus}`,
-      `目标：${result.objective || '只读探索'}`,
-      `摘要：${resultSummary}`,
-      `范围：${roots}`,
-      `查询：${queries}`,
-      `发现/读取：${filesDiscovered} / ${result.filesInspected} 个文件`,
-      duration ? `耗时：${duration}` : '',
-      ignoredPaths ? `忽略：${ignoredPaths}` : '',
-      stoppingCondition ? `停止条件：${stoppingCondition}` : '',
-      limitReasons ? `预算边界：${limitReasons}` : '',
-    ].filter((line) => line.length > 0).join('\n')
-    : '';
-  const processText = [
-    summaryText,
-    processLines.length > 0 ? `事件：${processLines.length}` : '',
-    processLines.join('\n'),
-  ].filter((line) => line.trim().length > 0).join('\n').trim();
-  const evidenceText = evidence.length > 0
-    ? [
-      `状态：${status}`,
-      `终态：${terminalStatus}`,
-      `目标：${result.objective || '只读探索'}`,
-      `证据：${evidence.length}`,
-      ...evidence.map((item) => [
-        `- ${item.path}${typeof item.line === 'number' ? `:${item.line}` : ''}`,
-        item.label,
-        typeof item.score === 'number' ? `分数 ${item.score}` : '',
-      ].filter(Boolean).join(' — ')),
-    ].join('\n')
-    : '';
-  const candidateText = candidateFiles.length > 0
-    ? [
-      `状态：${status}`,
-      `终态：${terminalStatus}`,
-      `目标：${result.objective || '只读探索'}`,
-      `发现/读取：${filesDiscovered} / ${result.filesInspected} 个文件`,
-      `候选：${candidateFiles.length}`,
-      ...candidateFiles.map((file) => [
-        `- ${file.path}`,
-        `分数 ${file.score}`,
-        file.reasons.length > 0 ? presentExploreAgentCandidateReasons(file.reasons) : '',
-      ].filter(Boolean).join(' — ')),
-    ].join('\n')
-    : '';
-  const matchesText = matches.length > 0
-    ? [
-      `状态：${status}`,
-      `终态：${terminalStatus}`,
-      `目标：${result.objective || '只读探索'}`,
-      `查询：${queries}`,
-      `命中片段：${matches.length}`,
-      ...matches.map((match) => `- ${match.path}:${match.line} [${match.query}] ${match.snippet}`),
-    ].join('\n')
-    : '';
-  const needsContinuation =
-    result.partial === true ||
-    !result.ok ||
-    Boolean(limitReasons) ||
-    result.terminalStatus === 'completed_empty';
-  const continuationReason = needsContinuation
-    ? presentExploreAgentContinuationReason({
-      partial: result.partial === true,
-      ok: result.ok,
-      hasLimitReasons: Boolean(limitReasons),
-      terminalStatus: result.terminalStatus,
-    })
-    : '';
-  const continuationText = needsContinuation
-    ? [
-      '继续这次只读探索，不要修改文件。',
-      continuationReason ? `续研原因：${continuationReason}` : '',
-      `上一轮状态：${status}`,
-      `上一轮终态：${terminalStatus}`,
-      `目标：${result.objective || '只读探索'}`,
-      `范围：${roots}`,
-      `查询：${queries}`,
-      `发现/读取：${filesDiscovered} / ${result.filesInspected} 个文件`,
-      duration ? `上一轮耗时：${duration}` : '',
-      ignoredPaths ? `继续忽略：${ignoredPaths}` : '',
-      stoppingCondition ? `停止条件：${stoppingCondition}` : '',
-      limitReasons ? `上一轮预算边界：${limitReasons}` : '',
-      resultSummary ? `上一轮摘要：${resultSummary}` : '',
-      candidateFiles.length > 0
-        ? [
-          '优先补读候选：',
-          ...candidateFiles.slice(0, 5).map((file) => `- ${file.path}（分数 ${file.score}）`),
-        ].join('\n')
-        : '',
-      matches.length > 0
-        ? [
-          '已有命中片段：',
-          ...matches.slice(0, 5).map((match) => `- ${match.path}:${match.line} [${match.query}] ${match.snippet}`),
-        ].join('\n')
-        : '',
-      '请只读检查仍缺证据的部分，输出新的证据锚点、候选文件、结论和下一步 gate。',
-    ].filter((line) => line.trim().length > 0).join('\n')
-    : '';
-
-  function copyButtonState(key: string, idleLabel: string, copiedAria: string) {
-    const phase = copyFeedback.phaseFor(key);
-    return {
-      phase,
-      disabled: copyFeedback.isPending,
-      label: phase === 'pending'
-        ? '复制中…'
-        : phase === 'copied'
-          ? '已复制'
-          : phase === 'failed'
-            ? '复制失败'
-            : idleLabel,
-      ariaLabel: phase === 'pending'
-        ? `${idleLabel}中`
-        : phase === 'copied'
-          ? copiedAria
-          : phase === 'failed'
-            ? `${idleLabel}失败`
-            : idleLabel,
-    };
-  }
-
-  const summaryCopy = copyButtonState('summary', '复制摘要', '已复制探索摘要');
-  const continuationCopy = copyButtonState('continuation', '复制续研提示', '已复制续研提示');
-  const processCopy = copyButtonState('process', '复制过程', '已复制探索过程');
-  const evidenceCopy = copyButtonState('evidence', '复制证据', '已复制证据锚点');
-  const reportCopy = copyButtonState('report', '复制报告', '已复制研究报告');
-  const candidateCopy = copyButtonState('candidate', '复制候选', '已复制候选文件');
-  const matchesCopy = copyButtonState('matches', '复制片段', '已复制命中片段');
-
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'agent' }))} data-kind="explore_agent" data-ok={result.ok ? 'true' : 'false'}>
-      <header className={previewVariants({ part: 'agent-head' })}>
-        <strong>{redactSecrets(result.objective || '只读探索')}</strong>
-        <small>
-          {status} · 发现/读 {filesDiscovered} / {result.filesInspected} 个文件 · {skippedSummary} · {formatBytes(result.bytesRead)}
-          {limitReasons ? ' · 受预算限制' : ''}
-          {continuationReason ? ` · 建议续研：${continuationReason}` : ''}
-          {duration ? ` · 耗时 ${duration}` : ''}
-        </small>
-        {resultSummary.length > 0 && (
-          <div className={previewVariants({ part: 'agent-summary-line' })}>
-            <small>{redactSecrets(resultSummary)}</small>
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('summary', summaryText)}
-              disabled={summaryCopy.disabled}
-              aria-label={summaryCopy.ariaLabel}
-              aria-busy={summaryCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={summaryCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={summaryCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={summaryCopy.phase === 'failed' ? 'true' : undefined}
-            >
-              {summaryCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{summaryCopy.label}</span>
-            </UiButton>
-          </div>
-        )}
-        {continuationText.length > 0 && (
-          <div className={previewVariants({ part: 'agent-actions' })} aria-label="只读探索后续操作">
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('continuation', continuationText)}
-              disabled={continuationCopy.disabled}
-              aria-label={continuationCopy.ariaLabel}
-              aria-busy={continuationCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={continuationCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={continuationCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={continuationCopy.phase === 'failed' ? 'true' : undefined}
-              title="复制一段可继续只读探索的提示"
-            >
-              {continuationCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{continuationCopy.label}</span>
-            </UiButton>
-          </div>
-        )}
-      </header>
-      {!result.ok && (
-        <div className={previewVariants({ part: 'agent-message' })} role="note">
-          {redactSecrets(result.message ?? '只读探索未完成。')}
-        </div>
-      )}
-      <dl className={previewVariants({ part: 'agent-meta' })}>
-        <div>
-          <dt>终态</dt>
-          <dd>{terminalStatus}</dd>
-        </div>
-        <div>
-          <dt>发现/读</dt>
-          <dd>{filesDiscovered} / {result.filesInspected} 个文件</dd>
-        </div>
-        <div>
-          <dt>范围</dt>
-          <dd>{redactSecrets(roots)}</dd>
-        </div>
-        <div>
-          <dt>查询</dt>
-          <dd>{redactSecrets(queries)}</dd>
-        </div>
-        {ignoredPaths && (
-          <div>
-            <dt>忽略</dt>
-            <dd>{redactSecrets(ignoredPaths)}</dd>
-          </div>
-        )}
-        {stoppingCondition && (
-          <div>
-            <dt>停止</dt>
-            <dd>{redactSecrets(stoppingCondition)}</dd>
-          </div>
-        )}
-        {limitReasons && (
-          <div>
-            <dt>边界</dt>
-            <dd>{redactSecrets(limitReasons)}</dd>
-          </div>
-        )}
-        {continuationReason && (
-          <div>
-            <dt>后续</dt>
-            <dd>建议续研：{redactSecrets(continuationReason)}</dd>
-          </div>
-        )}
-      </dl>
-      {progress.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="探索过程">
-          <div className={previewVariants({ part: 'agent-section-head' })}>
-            <strong>过程</strong>
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('process', processText)}
-              disabled={processCopy.disabled}
-              aria-label={processCopy.ariaLabel}
-              aria-busy={processCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={processCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={processCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={processCopy.phase === 'failed' ? 'true' : undefined}
-            >
-              {processCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{processCopy.label}</span>
-            </UiButton>
-          </div>
-          <ul>
-            {progress.map((item, index) => (
-              <li key={`${index}:${item.slice(0, 24)}`}>
-                <span>{redactSecrets(item)}</span>
-              </li>
-            ))}
-          </ul>
-        </section>
-      )}
-      {evidence.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="证据锚点">
-          <div className={previewVariants({ part: 'agent-section-head' })}>
-            <strong>证据锚点</strong>
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('evidence', evidenceText)}
-              disabled={evidenceCopy.disabled}
-              aria-label={evidenceCopy.ariaLabel}
-              aria-busy={evidenceCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={evidenceCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={evidenceCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={evidenceCopy.phase === 'failed' ? 'true' : undefined}
-            >
-              {evidenceCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{evidenceCopy.label}</span>
-            </UiButton>
-          </div>
-          <ul>
-            {evidence.map((item, index) => (
-              <li key={`${item.path}:${item.line ?? 'file'}:${index}`}>
-                <code>
-                  {redactSecrets(item.path)}
-                  {typeof item.line === 'number' ? `:${item.line}` : ''}
-                </code>
-                <small>
-                  {redactSecrets(item.label)}
-                  {typeof item.score === 'number' ? ` · 分数 ${item.score}` : ''}
-                </small>
-              </li>
-            ))}
-          </ul>
-        </section>
-      )}
-      {reportLines.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="研究报告">
-          <div className={previewVariants({ part: 'agent-section-head' })}>
-            <strong>研究报告</strong>
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('report', reportText)}
-              disabled={reportCopy.disabled}
-              aria-label={reportCopy.ariaLabel}
-              aria-busy={reportCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={reportCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={reportCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={reportCopy.phase === 'failed' ? 'true' : undefined}
-            >
-              {reportCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{reportCopy.label}</span>
-            </UiButton>
-          </div>
-          <ul>
-            {reportLines.map((line, index) => (
-              <li key={`${index}:${line.slice(0, 24)}`}>
-                <span>{redactSecrets(line)}</span>
-              </li>
-            ))}
-          </ul>
-        </section>
-      )}
-      {candidateFiles.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="候选文件">
-          <div className={previewVariants({ part: 'agent-section-head' })}>
-            <strong>候选文件</strong>
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('candidate', candidateText)}
-              disabled={candidateCopy.disabled}
-              aria-label={candidateCopy.ariaLabel}
-              aria-busy={candidateCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={candidateCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={candidateCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={candidateCopy.phase === 'failed' ? 'true' : undefined}
-            >
-              {candidateCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{candidateCopy.label}</span>
-            </UiButton>
-          </div>
-          <ul>
-            {candidateFiles.map((file) => (
-              <li key={file.path}>
-                <code>{redactSecrets(file.path)}</code>
-                <small>
-                  分数 {file.score}
-                  {file.reasons.length > 0 ? ` · ${presentExploreAgentCandidateReasons(file.reasons)}` : ''}
-                </small>
-              </li>
-            ))}
-          </ul>
-        </section>
-      )}
-      {matches.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="命中片段">
-          <div className={previewVariants({ part: 'agent-section-head' })}>
-            <strong>命中片段</strong>
-            <UiButton
-              type="button"
-              variant="ghost"
-              size="sm"
-              className={previewVariants({ part: 'agent-copy' })}
-              onClick={() => void copyFeedback.copy('matches', matchesText)}
-              disabled={matchesCopy.disabled}
-              aria-label={matchesCopy.ariaLabel}
-              aria-busy={matchesCopy.phase === 'pending' ? 'true' : undefined}
-              data-pending={matchesCopy.phase === 'pending' ? 'true' : undefined}
-              data-copied={matchesCopy.phase === 'copied' ? 'true' : 'false'}
-              data-copy-error={matchesCopy.phase === 'failed' ? 'true' : undefined}
-            >
-              {matchesCopy.phase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-              <span>{matchesCopy.label}</span>
-            </UiButton>
-          </div>
-          <ul>
-            {matches.map((match, index) => (
-              <li key={`${match.path}:${match.line}:${index}`}>
-                <code>{redactSecrets(match.path)}:{match.line}</code>
-                <small>{redactSecrets(match.query)}</small>
-                <p>{redactSecrets(match.snippet)}</p>
-              </li>
-            ))}
-          </ul>
-        </section>
-      )}
-      {notes.length > 0 && (
-        <section className={previewVariants({ part: 'agent-section' })} aria-label="探索说明">
-          <strong>说明</strong>
-          <ul>
-            {notes.map((note, index) => (
-              <li key={`${index}:${note.slice(0, 24)}`}>
-                <span>{redactSecrets(note)}</span>
-              </li>
-            ))}
-          </ul>
-        </section>
-      )}
-    </div>
-  );
-}
-
-function presentExploreAgentTerminalStatus(
-  terminalStatus: Extract<ToolResultContent, { kind: 'explore_agent' }>['terminalStatus'],
-  ok: boolean,
-  partial: boolean,
-  reason: Extract<ToolResultContent, { kind: 'explore_agent' }>['reason'],
-): string {
-  switch (terminalStatus) {
-    case 'completed':
-      return '完成，有证据';
-    case 'completed_empty':
-      return '完成，无证据';
-    case 'failed':
-      return '失败';
-    case 'canceled':
-      return '已取消';
-    case 'canceled_partial':
-      return '已取消，有部分结果';
-    case undefined:
-      if (reason === 'aborted' && partial) return '已取消，有部分结果';
-      if (reason === 'aborted') return '已取消';
-      if (!ok) return '失败';
-      return '完成';
-    default:
-      return '未知终态';
-  }
-}
-
-function presentExploreAgentReason(
-  reason: Extract<ToolResultContent, { kind: 'explore_agent' }>['reason'],
-): string | undefined {
-  switch (reason) {
-    case 'invalid_objective':
-      return '目标无效';
-    case 'invalid_root':
-      return '范围无效';
-    case 'no_readable_roots':
-      return '没有可读取范围';
-    case 'aborted':
-      return '已取消';
-    case undefined:
-      return undefined;
-    default:
-      return '未知诊断';
-  }
-}
-
-function presentExploreAgentLimitReason(reason: string): string {
-  switch (reason) {
-    case 'candidate_budget':
-      return '候选文件预算已满';
-    case 'file_budget':
-      return '读取文件预算已满';
-    case 'match_budget':
-      return '命中预算已满';
-    case 'byte_budget':
-      return '读取字节预算已满';
-    default:
-      return '';
-  }
-}
-
-function presentExploreAgentContinuationReason(input: {
-  partial: boolean;
-  ok: boolean;
-  hasLimitReasons: boolean;
-  terminalStatus: Extract<ToolResultContent, { kind: 'explore_agent' }>['terminalStatus'];
-}): string {
-  if (input.partial) return '已有部分结果，仍需补证据';
-  if (!input.ok) return '上一轮未完成';
-  if (input.hasLimitReasons) return '达到预算边界';
-  if (input.terminalStatus === 'completed_empty') return '没有找到证据';
-  return '仍缺证据';
-}
-
-function formatExploreAgentEvent(event: { type: string; message: string; at?: number }, startedAt?: number): string {
-  const label = presentExploreAgentEventType(event.type);
-  const message = typeof event.message === 'string' ? event.message.trim() : '';
-  const offset = formatExploreAgentEventOffset(event.at, startedAt);
-  const prefix = [label, offset].filter(Boolean).join(' ');
-  return prefix ? `${prefix}：${message}` : message;
-}
-
-function formatExploreAgentEventOffset(at: number | undefined, startedAt: number | undefined): string {
-  if (typeof at !== 'number' || typeof startedAt !== 'number') return '';
-  if (!Number.isFinite(at) || !Number.isFinite(startedAt)) return '';
-  const delta = Math.max(0, Math.floor(at - startedAt));
-  const formatted = formatDuration(delta);
-  return formatted ? `+${formatted}` : '';
-}
-
-function presentExploreAgentEventType(type: string): string {
-  switch (type) {
-    case 'started':
-      return '开始';
-    case 'scope_resolved':
-      return '范围';
-    case 'scan':
-      return '扫描';
-    case 'read':
-      return '读取';
-    case 'checkpoint':
-      return '进度';
-    case 'completed':
-      return '完成';
-    case 'failed':
-      return '失败';
-    case 'aborted':
-      return '取消';
-    default:
-      return '';
-  }
-}
-
-function presentExploreAgentCandidateReasons(reasons: string[]): string {
-  return reasons.map((reason) => {
-    if (reason === 'content match') return '内容命中';
-    if (reason === 'project manifest') return '项目配置';
-    if (reason === 'project documentation') return '项目文档';
-    if (reason === 'project entrypoint') return '入口文件';
-    if (reason === 'project test surface') return '测试线索';
-    if (reason === 'project source surface') return '源码线索';
-    const pathMatch = reason.match(/^path contains "(.+)"$/);
-    if (pathMatch) return `路径命中 ${redactSecrets(pathMatch[1] ?? '')}`;
-    return '探索线索';
-  }).join(', ');
-}
-
-/* PR-FORMAT-BYTES-DEDUP-0 (round 21/30): made exported so
-   `apps/desktop/src/renderer/artifact-pane.tsx` can drop its
-   local duplicate. Standardized on KB/MB labels (artifact-pane
-   was already user-visible with KB/MB) plus the robust
-   non-finite/<=0 guard from the previous components.tsx form.
-   Both consumers now produce identical output for the same
-   `bytes` input. */
-export function formatBytes(bytes: number): string {
-  if (!Number.isFinite(bytes) || bytes <= 0) return '0 B';
-  if (bytes < 1024) return `${Math.round(bytes)} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-function OfficeDocumentPreview(props: {
-  result: Extract<ToolResultContent, { kind: 'office_document' }>;
-}) {
-  const { result } = props;
-  const stdout = capLines(redactSecrets(result.stdout ?? ''));
-  const stderr = capLines(redactSecrets(result.stderr ?? ''));
-  const message = result.message ? redactSecrets(result.message) : '';
-  const args = result.args?.map((arg) => redactSecrets(arg)).join(' ');
-  const title = result.path ? redactSecrets(result.path) : 'Office 文档';
-  const operation = result.operation ? redactSecrets(result.operation) : '未执行';
-  const reason = presentOfficeDocumentReason(result.reason);
-  const hasOutput = stdout.body.length > 0 || stderr.body.length > 0;
-
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'office' }))} data-kind="office_document" data-ok={result.ok ? 'true' : 'false'}>
-      <header className={previewVariants({ part: 'office-head' })}>
-        <strong>{title}</strong>
-        <small>
-          {operation}
-          {result.ok ? ' · 已完成' : ' · 未完成'}
-          {result.truncated ? ' · 输出已截断' : ''}
-        </small>
-      </header>
-      {args && <code className={previewVariants({ part: 'office-args' })}>officecli {args}</code>}
-      {!result.ok && (
-        <div className={previewVariants({ part: 'office-message' })} role="note">
-          <span>{message || 'Office 文档操作未完成。'}</span>
-          {reason && <small>诊断：{reason}</small>}
-        </div>
-      )}
-      {result.ok && !hasOutput && <p className={previewVariants({ part: 'office-empty' })}>（无输出）</p>}
-      {stdout.body.length > 0 && (
-        <pre className={previewVariants({ part: 'office-stream' })} data-stream="stdout">
-          {stdout.body}
-          {stdout.capped > 0 && `\n\n… stdout 已隐藏 ${stdout.capped} 行`}
-        </pre>
-      )}
-      {stderr.body.length > 0 && (
-        <pre className={previewVariants({ part: 'office-stream' })} data-stream="stderr">
-          {stderr.body}
-          {stderr.capped > 0 && `\n\n… stderr 已隐藏 ${stderr.capped} 行`}
-        </pre>
-      )}
-    </div>
-  );
-}
-
-function presentOfficeDocumentReason(reason: string | undefined): string | undefined {
-  switch (reason) {
-    case 'invalid_operation':
-      return '操作不支持';
-    case 'invalid_path':
-      return '路径无效';
-    case 'unsupported_extension':
-      return '文件类型不支持';
-    case 'missing_file':
-      return '文件不存在';
-    case 'not_file':
-      return '不是文件';
-    case 'symlink_escape':
-      return '符号链接被拒绝';
-    case 'invalid_selector':
-      return '选择器无效';
-    case 'invalid_query':
-      return '查询表达式无效';
-    case 'invalid_props':
-      return '属性无效';
-    case 'file_exists':
-      return '文件已存在';
-    case 'officecli_missing':
-      return 'officecli 未安装';
-    case 'officecli_timeout':
-      return '操作超时';
-    case 'officecli_failed':
-      return '操作失败';
-    case undefined:
-      return undefined;
-    default:
-      return '未知诊断';
-  }
-}
-
-/**
- * Line-level diff coloring. Splits the unified-diff text on newlines and
- * tags each line with `data-line="add" | "del" | "hunk" | "meta" | "ctx"`
- * for CSS to color. Doesn't try to parse the hunk semantics — we leave
- * that to a future inline editor view; this is just a readable preview.
- */
-/**
- * PR-CHAT-WEB-SEARCH-RENDER-0 — plain-text card list for the gated
- * WebSearch agent tool result. Matches the Settings → 联网搜索 live-query
- * verification layout so the user gets the same shape whether the search came
- * from a manual verification run or the agent. Never renders markdown / HTML;
- * each cell is `redactSecrets`'d as a belt-and-braces guard against
- * a provider response that happened to echo a token.
- */
-function WebSearchPreview(props: {
-  query: string;
-  provider: string;
-  rows: ReadonlyArray<{ title: string; url: string; snippet: string; source: string }>;
-}) {
-  const rows = props.rows
-    .map((row) => {
-      const normalizedUrl = normalizeSearchUrl(row.url);
-      if (!normalizedUrl.ok) return null;
-      return { ...row, url: redactSecrets(normalizedUrl.value) };
-    })
-    .filter((row): row is { title: string; url: string; snippet: string; source: string } => row !== null);
-
-  if (rows.length === 0) {
-    return (
-      <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'web-search' }))} data-kind="web_search">
-        <header>
-          <strong>{redactSecrets(props.query)}</strong>
-          <small>{props.provider} · 没有结果</small>
-        </header>
-      </div>
-    );
-  }
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'web-search' }))} data-kind="web_search">
-      <header>
-        <strong>{redactSecrets(props.query)}</strong>
-        <small>
-          {props.provider} · {rows.length} 条结果
-        </small>
-      </header>
-      <ul>
-        {rows.map((row, idx) => (
-          <li key={`${row.url}-${idx}`}>
-            <a href={row.url} target="_blank" rel="noreferrer noopener">
-              {redactSecrets(row.title)}
-            </a>
-            <small>{redactSecrets(row.source)}</small>
-            <p>{redactSecrets(row.snippet)}</p>
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-}
-
-function WebSearchErrorPreview(props: {
-  query?: string;
-  provider: string;
-  reason: string;
-  message: string;
-  credentialSource?: string;
-}) {
-  const sourceCopy =
-    props.credentialSource === 'env'
-      ? '环境变量'
-      : props.credentialSource === 'saved'
-        ? '本机已保存 key'
-        : props.credentialSource === 'none'
-          ? '未配置'
-          : '来源未知';
-  const repairCopy =
-    props.reason === 'invalid_credentials' && props.credentialSource === 'env'
-      ? '请检查 TAVILY_API_KEY / MAKA_TAVILY_API_KEY 后重启。'
-      : props.reason === 'invalid_credentials'
-        ? '请在 设置 · 联网搜索 中更新 Tavily key。'
-        : props.reason === 'rate_limited'
-          ? 'Tavily 当前限流，请稍后重试或更换可用凭据。'
-          : props.reason === 'not_configured'
-            ? '请先完成联网搜索配置后再重试。'
-            : props.reason === 'timeout'
-              ? '请求超时，请稍后重试。'
-              : props.reason === 'incognito_active'
-                ? '隐私模式下不会发起联网搜索。'
-                : '请检查网络或稍后重试。';
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'web-search' }), previewVariants({ part: 'web-search-error' }))} data-kind="web_search_error">
-      <header>
-        <strong>{redactSecrets(props.query ?? '联网搜索')}</strong>
-        <small>{redactSecrets(props.provider)} · 搜索失败 · {sourceCopy}</small>
-      </header>
-      <p className={previewVariants({ part: 'web-search-error-message' })}>{redactSecrets(props.message)}</p>
-      <p className={previewVariants({ part: 'web-search-error-repair' })}>{repairCopy}</p>
-    </div>
-  );
-}
-
-function FileDiffPreview(props: { diff: string; paths: string[] }) {
-  // Apply UI-level redaction then cap the displayed lines. Both are
-  // @kenji's PR76 review items: never echo a token a tool happened to dump
-  // into a diff (commit body, .env file diff, etc.), and never let a
-  // 10k-line diff create 10k React elements.
-  const { body, capped } = capLines(redactSecrets(props.diff));
-  const lines = body.split('\n');
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'diff' }))} data-kind="file_diff">
-      {props.paths.length > 0 && (
-        <div className={previewVariants({ part: 'diff-paths' })}>
-          {props.paths.map((path) => (
-            <code key={path}>{path}</code>
-          ))}
-        </div>
-      )}
-      <pre className={previewVariants({ part: 'diff-body' })}>
-        {lines.map((line, index) => (
-          <span
-            key={`${index}:${line.slice(0, 16)}`}
-            className={previewVariants({ part: 'diff-line' })}
-            data-line={diffLineKind(line)}
-          >
-            {line || ' '}
-            {'\n'}
-          </span>
-        ))}
-        {capped > 0 && (
-          <span className={previewVariants({ part: 'diff-line' })} data-line="meta">
-            {`\n… 已隐藏 ${capped} 行\n`}
-          </span>
-        )}
-      </pre>
-    </div>
-  );
-}
-
-function diffLineKind(line: string): 'add' | 'del' | 'hunk' | 'meta' | 'ctx' {
-  if (line.startsWith('+++') || line.startsWith('---')) return 'meta';
-  if (line.startsWith('@@')) return 'hunk';
-  if (line.startsWith('+')) return 'add';
-  if (line.startsWith('-')) return 'del';
-  return 'ctx';
-}
-
-/**
- * Terminal output preview. Shows the command + working directory header,
- * an exit-code badge tinted by success/failure, then stdout and stderr
- * in separate blocks (stderr only rendered when non-empty, in destructive
- * tone). Empty output gets an explicit "(no output)" placeholder so a
- * silent successful command doesn't look like a render bug.
- */
-function TerminalPreview(props: {
-  cwd: string;
-  cmd: string;
-  exitCode: number;
-  stdout: string;
-  stderr: string;
-}) {
-  const copyFeedback = useClipboardCopyFeedback();
-  const succeeded = props.exitCode === 0;
-  const hasOutput = props.stdout.length > 0 || props.stderr.length > 0;
-  // Redact + cap stdout/stderr independently. `npm test` against a misconfigured
-  // provider can dump megabytes of stderr; we keep the first TOOL_LINE_CAP
-  // lines and append a hidden-count marker.
-  const stdout = capLines(redactSecrets(props.stdout));
-  const stderr = capLines(redactSecrets(props.stderr));
-  // The cmd line is also user-runtime text — don't echo a `--api-key=...`
-  // arg into the chat without masking it.
-  const safeCmd = redactSecrets(props.cmd);
-  const safeCwd = redactSecrets(props.cwd);
-  const hiddenLines = stdout.capped + stderr.capped;
-  const handoffText = [
-    '终端输出需要继续研读',
-    `工作目录：${safeCwd}`,
-    `命令：${safeCmd}`,
-    `退出码：${props.exitCode}`,
-    `截断：stdout 已隐藏 ${stdout.capped} 行，stderr 已隐藏 ${stderr.capped} 行`,
-    stdout.body.length > 0 ? `stdout 预览：\n${stdout.body}` : '',
-    stderr.body.length > 0 ? `stderr 预览：\n${stderr.body}` : '',
-    '请在深度研究 / 只读探索里结合相关路径确认完整输出影响和下一步。',
-  ].filter((line) => line.length > 0).join('\n\n');
-
-  const handoffCopyPhase = copyFeedback.phaseFor('handoff');
-  const handoffCopyLabel = handoffCopyPhase === 'pending'
-    ? '复制中…'
-    : handoffCopyPhase === 'copied'
-      ? '已复制'
-      : handoffCopyPhase === 'failed'
-        ? '复制失败'
-        : '复制研读提示';
-  const handoffCopyAria = handoffCopyPhase === 'pending'
-    ? '复制终端研读提示中'
-    : handoffCopyPhase === 'copied'
-      ? '已复制终端研读提示'
-      : handoffCopyPhase === 'failed'
-        ? '复制终端研读提示失败'
-        : '复制终端研读提示';
-
-  return (
-    <div className={cn(previewVariants({ part: 'overlay' }), previewVariants({ part: 'terminal' }))} data-kind="terminal">
-      <header className={previewVariants({ part: 'terminal-head' })}>
-        <code className={previewVariants({ part: 'terminal-cwd' })}>{safeCwd}</code>
-        <code className={previewVariants({ part: 'terminal-cmd' })}>$ {safeCmd}</code>
-        <span
-          className={previewVariants({ part: 'terminal-exit' })}
-          data-ok={succeeded ? 'true' : 'false'}
-          aria-label={`退出码 ${props.exitCode}`}
-        >
-          退出码 {props.exitCode}
-        </span>
-      </header>
-      {!hasOutput && <p className={previewVariants({ part: 'terminal-empty' })}>（无输出）</p>}
-      {props.stdout.length > 0 && (
-        <pre className={previewVariants({ part: 'terminal-stream' })} data-stream="stdout">
-          {stdout.body}
-          {stdout.capped > 0 && `\n\n… stdout 已隐藏 ${stdout.capped} 行`}
-        </pre>
-      )}
-      {props.stderr.length > 0 && (
-        <pre className={previewVariants({ part: 'terminal-stream' })} data-stream="stderr">
-          {stderr.body}
-          {stderr.capped > 0 && `\n\n… stderr 已隐藏 ${stderr.capped} 行`}
-        </pre>
-      )}
-      {hiddenLines > 0 && (
-        <div className={previewVariants({ part: 'terminal-truncated-note' })}>
-          <span>
-            输出较长，当前只展示每路输出的前 {TOOL_LINE_CAP} 行。需要继续研读时，可以切到深度研究并把命令、相关路径和想确认的问题交给只读探索。
-          </span>
-          <PrimitiveButton
-            type="button"
-            variant="ghost"
-            size="sm"
-            className={previewVariants({ part: 'terminal-copy' })}
-            onClick={() => void copyFeedback.copy('handoff', handoffText)}
-            disabled={handoffCopyPhase === 'pending'}
-            aria-label={handoffCopyAria}
-            aria-busy={handoffCopyPhase === 'pending' ? 'true' : undefined}
-            data-pending={handoffCopyPhase === 'pending' ? 'true' : undefined}
-            data-copied={handoffCopyPhase === 'copied' ? 'true' : 'false'}
-            data-copy-error={handoffCopyPhase === 'failed' ? 'true' : undefined}
-          >
-            {handoffCopyPhase === 'copied' ? <Check size={13} strokeWidth={2} aria-hidden="true" /> : <Copy size={13} strokeWidth={1.75} aria-hidden="true" />}
-            <span>{handoffCopyLabel}</span>
-          </PrimitiveButton>
-        </div>
-      )}
-    </div>
-  );
-}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 function mergeTools(stored: ToolActivityItem[], live: ToolActivityItem[]): ToolActivityItem[] {
   const byId = new Map(stored.map((item) => [item.toolUseId, item]));
@@ -7893,70 +4371,3 @@ function mergeTools(stored: ToolActivityItem[], live: ToolActivityItem[]): ToolA
 }
 
 const noMessagesYet = '暂无消息';
-
-interface SessionGroup {
-  label: string;
-  sessions: SessionSummary[];
-}
-
-/**
- * In the Chats filter, pinned (flagged) sessions float to the top in their
- * own section per the session-list-lifecycle contract, separate from the
- * date-bucketed remainder. Other filters keep the date-bucket layout.
- */
-function groupSessionsForFilter(sessions: SessionSummary[], selection: NavSelection): SessionGroup[] {
-  if (selection.section !== 'sessions' || selection.filter !== 'chats') {
-    return groupSessionsByTime(sessions);
-  }
-  const pinned = sessions.filter((session) => session.isFlagged);
-  const rest = sessions.filter((session) => !session.isFlagged);
-  const groups: SessionGroup[] = [];
-  if (pinned.length > 0) {
-    groups.push({ label: '已置顶', sessions: pinned });
-  }
-  return [...groups, ...groupSessionsByTime(rest)];
-}
-
-/**
- * Cluster the session list into Today / Yesterday / Past 7 days / Past 30 days
- * / Older buckets. Sorted by lastMessageAt descending within each group. Falls
- * back to a single bucket if every session lacks a timestamp.
- */
-function groupSessionsByTime(sessions: SessionSummary[]): SessionGroup[] {
-  const now = Date.now();
-  const startOfToday = new Date(now);
-  startOfToday.setHours(0, 0, 0, 0);
-  const todayMs = startOfToday.getTime();
-  const yesterdayMs = todayMs - 24 * 60 * 60 * 1000;
-  const sevenDaysMs = todayMs - 7 * 24 * 60 * 60 * 1000;
-  const thirtyDaysMs = todayMs - 30 * 24 * 60 * 60 * 1000;
-
-  const buckets: SessionGroup[] = [
-    { label: '今天', sessions: [] },
-    { label: '昨天', sessions: [] },
-    { label: '过去 7 天', sessions: [] },
-    { label: '过去 30 天', sessions: [] },
-    { label: '更早', sessions: [] },
-    { label: '待发送', sessions: [] },
-  ];
-
-  for (const session of sessions) {
-    const at = session.lastMessageAt;
-    if (!at) {
-      buckets[5]!.sessions.push(session);
-      continue;
-    }
-    if (at >= todayMs) buckets[0]!.sessions.push(session);
-    else if (at >= yesterdayMs) buckets[1]!.sessions.push(session);
-    else if (at >= sevenDaysMs) buckets[2]!.sessions.push(session);
-    else if (at >= thirtyDaysMs) buckets[3]!.sessions.push(session);
-    else buckets[4]!.sessions.push(session);
-  }
-
-  return buckets.filter((group) => group.sessions.length > 0);
-}
-
-function formatSessionMeta(session: SessionSummary): string {
-  if (!session.lastMessageAt) return noMessagesYet;
-  return formatRelativeTimestamp(session.lastMessageAt);
-}
