+++ packages/ui/src/nav-selection.ts
@@ -1,0 +1,7 @@
+export type SessionFilter = 'chats' | 'flagged' | 'archived';
+
+export type NavSelection =
+  | { section: 'sessions'; filter: SessionFilter }
+  | { section: 'automations' }
+  | { section: 'skills' }
+  | { section: 'daily-review' };
